[roll] Roll fuchsia [microbenchmarks] Make HandleValid test's set of metric names deterministic
For the allowlist being added for performance metric names, it's
desirable for the set of metric names to be deterministic.

Before this change, the set of metric names would vary based on the
number of CPUs on the target system.  Address that by always testing
two thread counts (one thread, and one thread per CPU).

Note that the NUCs we test on only have two CPUs anyway (with
hyperthreading disabled, which is how we run the performance tests).

Original-Bug: 105202
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/709286
Original-Revision: 1e1be75750ab86bd64cffac43b7014c8c4e8f9b1
GitOrigin-RevId: adf9065e962d0b384caedefdcd53b96c3fcfb385
Change-Id: I042ccb2949309fba199b167991504aecd508a02a