[roll] Roll fuchsia [flatland] Update vulkan image pipe swapchain to Flatland
This CL updates vulkan image pipe swapchain to use Flatland instead of
the Gfx ImagePipe which is being deprecated. We do not change the
signature or the name of the extension. Instead we reinterpret zx_handle
such that it is a ViewCreation token.

Original-Bug: 72562
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/694188
Original-Revision: 894e67077cfe20a742616f9029ecd68f5e5766b5
GitOrigin-RevId: 7c6a411e4cc6a3229ec0c46a9b197284ae04d5dd
Change-Id: I7bdd25c269b127c8aa96f4e98cd188d7a5c5fda5