[roll] Roll fuchsia [diagnostics] Rename triage lib to avoid conflict
Right now the existence of a triage library and a triage binary causes
generated cargo files to have a broken dependency tree in some cases
(where they depend on both the executable library and the lib library).
This renames the triage lib so it doesn't get confused with the binary
target anymore.

Found this while attempting to gen-cargo the ffx target, but the
dependency that creates this problem is in triage_lib_test.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/709284
Original-Revision: 3b2bbb7c41a507535011cac08dc8aafc4fe48967
GitOrigin-RevId: 1b3d6e9c25d7f2f2679ec2cfa6d05807d0515d96
Change-Id: I77b9739bd53a586002fbcb6b2de6fa7530e62b54