[roll] Roll fuchsia [netstack3] Add routes for multicast subnets
When a new device is added, add routes for the IPv4 and IPv6 multicast
subnet. Without this, attempting to send multicast packets on a UDP
socket returns a "no route" error. With this, multicast packets are sent
as expected.

Original-Bug: 39479
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/709584
Original-Revision: 4835d04f724361bc6c4b20b38364022ea82b65c2
GitOrigin-RevId: 6fee2a00243dad96a3041abc252c169926eeee91
Change-Id: I496c00d1f5cf63e1e017342849b6b53c1de643e8