[roll] Roll fuchsia [local_storage] Change logged metric type to Integer for two metrics.
These two metrics are incorrectly being logged as OCCURRENCE metric
type, when they are INTEGER metric types in the Cobalt registry.
This was caused by the migration from Cobalt 1.1, due to the
replacement metric being of type INTEGER when the original 1.0
metric was of type EVENT_COUNT (which in Lapis is an Occurrence
metric type). See the bug for more details on this issue.

Original-Bug: 106179
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/709768
Original-Revision: 3e270e0e666d40b47711d654f6d7baa00390a826
GitOrigin-RevId: 7bfaf66fb9e0dfe2db2f8b32e080aa38aa33e0dd
Change-Id: I48b18fb4e4f5ed59bcc99bdd4c53c732b50fdf28