[roll] Roll fuchsia [netstack3] Support addr scope for UDP connect
Support scoped IP addresses for UDP socket connect operations. This
allows applications to provide a link-local address with a device
identifier as a remote address, or to bind the socket to a device and
provide a link-local address without a scope.

Original-Bug: 46566
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/707903
Original-Revision: b71a1cfd6747f28da151eec445a3a50ecb4986f5
GitOrigin-RevId: bf359b89d17a0de3087cd321497a163022f4b5b8
Change-Id: I65833fe1f144a9c14901be5adef52036159b1c5f