[roll] Roll fuchsia [settings] Update config tests to verify light group names have unique hashes
Light group names in the light hardware configuration are expected to be
unique and also to have unique hashes in order for the light setting to
function properly. This CL adds a test that panics if two light group
names hash to the same value.

Original-Fixed: 103000
Test: fx test sample-setui-config-test
      locally tested with two identical light group names and verified
      the test fails. Unfortunately I don't know of two strings that
      hash to the same value in Rust.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/706592
Original-Revision: ed8c996b8993ca39b3fd4b8dda718235e7002707
GitOrigin-RevId: b8554d16074fc62b9ed273f8a13701011c02a1a2
Change-Id: I59a7ec796e2a8cbebab69b30a1e979f630a39dfb