[roll] Roll fuchsia [ffx][target] Format IpV6 Address in TargetAddr
Update Display for TargetAddr to write the IPV6 addresses with
surrounding square brackets "[]".

Also if the address is a local address and has a scope, then it will translate
the scope id to its associated interface name (if possible) and append
that to the address.

Original-Bug: 104635
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/707970
Original-Revision: 99546363d2431fb144009b73a6d99cb6b0953a1f
GitOrigin-RevId: 42e23d5ed07b8bfebe68e2d65597f9b75616cabc
Change-Id: I38426bb1f3457f6c54c49f8f84bc523ade5325cf