[roll] Roll fuchsia [pbm] Support PBMs with correct relative paths
bundle_fetcher now supports PBMs with the new relative paths, and falls
back to the old format if those files are not found. We have to complete
this in a soft-transition, because bundle_fetcher is pinned to a
specific version in infra.

Original-Bug: 106117
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/709886
Original-Revision: 05973ebac86c2bb277ffcbaa5f2dade06bbf52fd
GitOrigin-RevId: 28dd1ec895013d5ff2aaecee1c0bd23a8e813b30
Change-Id: Iac2cf62b37386629f56c07c706bbaa98dd787ad0