[roll] Roll fuchsia [fxfs] Manage volumes explicitly on the client
Lifts the responsibility of creating the default volume to the client.
For now, we bake it into fs_management and fshost to always do this, so
this is generally a pure refactoring in terms of external behaviours.

That said, this involves a sweeping change to fs_management and all
clients thereof.  This is centered around the introduction of volume
management to fs_management, via:
- StartedMultiVolumeFilesystem in C++, and
- ServingMultiVoluemFilesystem in Rust.

A few other changes are made:
- migrate_root is no longer used and can be removed in a followup CL.
- Fxfs no longer implements request queueing, this is done by
  fshost anyways.
- Some minor adjustments are made to the export root layout, e.g.
  volumes export a svc/ directory.
- A number of minor bugs were fixed, e.g. swallowing errors during
  start (cf. fxrev.dev/702516) and the fs-management minfs tests
  not actually launching a componentized minfs when they say they do.

Original-Bug: 93066
Original-Bug: 104313

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/700417
Size-Review: Gopi Krishna Chitluri <gopichitluri@google.com>
Original-Revision: 4c922be284e69d53d2792560c1b006e296fb4c55
GitOrigin-RevId: 3413bd6052f3a7e1fffa748494d0a31401d3a9dc
Change-Id: I074f9c11ef5cdd82879763ab6cf953ff0f68ecaf