[roll] Roll fuchsia [fidlgen][cpp] Introduce DeclDepGraph
This change introduces an abstraction for computing the dependency
graph among a library's declarations. It will allow use to remove
`declaration_order` from the IR - and, as a demonstration of that, we
migrate the C++ bindings over to relying upon this instead in the
following change.

We also fix a minor fidlc bug that was generating the wrong JSON
"maybe_size" element for layout members (on which this CL introduces a
dependency).

Original-Bug: 7660, 51002
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/703884
Original-Revision: 7062db5408806fbbfe2161eb28027a2bec85c0d0
GitOrigin-RevId: eb86f2be2c6ac1901a79dc776faf434a2bd945d7
Change-Id: I98895997183708f29da9f2186f99f8e3ecbbda98