[roll] Roll fuchsia [netstack3] Combine multicast interface selectors
Use a single type to represent the possible ways to select a multicast
interface when making a membership change. This makes it impossible to
represent the case where an interface ID and local address are provided,
pushing the responsibility of deciding what that means up to the caller.

Original-Bug: 39479
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/710224
Original-Revision: 49c4f6edbec2c67f693ab075a8efe71ada2a44c9
GitOrigin-RevId: 40dfc3d84ea73ccc235e3358b959c51abb625eb1
Change-Id: Ib9472ca7973726a81c9ee0551c059d229a8bd49d