[roll] Roll fuchsia [virtualization] Re-enable virtio-gpu tests
These integration tests had been disabled and had acquired some bitrot
because they had not been building in CI for some time. This CL revives
the binary but keeps many of the pre-existing tests as disabled. The
tests are now configured to run in a UITestManager environment that has
access to a hermetic graphics stack.

We introduce a new `virtio_gpu_test_util` binary that runs in a
(currently linux-only) guest and uses the linux Framebuffer API to
detect display geometry. We add a new test that simply compares that the
guest display geometry matches that of the Fuchsia view in the hermetic
UI stack.

Original-Bug: 102870
Test: fx test virtualization-gpu-tests --count 100
Multiply: virtualization-gpu-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/708566
Original-Revision: f71876f054db5b8b0c660feaa746fe47e844b633
GitOrigin-RevId: 3c52d11ab88f13ea5cf528f949a43f9c827579e1
Change-Id: Ib757c608f649a8b815dfef10deb12549c2f59c17