[roll] Roll fuchsia [zxdb][DAP] Initialize Request returns failure if not connected
This changes the DAP handler for the InitializeRequest message
so that if the debugger is not connected to the device, an error
is returned indicating the initialization failed and the debug
session should not continue.

The error message from the connection attempt is stored in the
session and set when the connection attempt is completed.

Original-Fixed: 105906
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/709703
Original-Revision: 9ff2f76f36fce23a4dad814b07e75c6c47a6fef1
GitOrigin-RevId: f1155e505273396d55117c0cb7e5d73e6470e35d
Change-Id: I5daea7dc4bd9bda5611b6a3193fc6591c4745f20