[roll] Roll fuchsia [cmc] Remove restricted feature for structured config.
We have completed all of our plans for structured config changes
that would be difficult to soft-migrate, so it's now pretty safe
for OOT integrations to be built.

This change does not on its own make it possible for OOT usage,
but the build integrations we will now write will no longer need
to opt-in to an unstable feature.

Original-Bug: 95368
Original-Bug: 101970
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/710404
Original-Revision: db34dcc8924090973f7090150fca70599119142a
GitOrigin-RevId: dfd97549638e1538bd3f2f92495af4fd76f58e09
Change-Id: I0d0f6157e700baf629f12d8d3b4eeedefb1008c2