[roll] Roll fuchsia [sysmem] sysmem-unittest migration fixup
We were running sysmem-bootfs-test but not sysmem-test,
so the migration of sysmem-test needed a couple tweaks.

To avoid this sort of regression in future, go ahead
and run sysmem-test as well.  The actual tests are the
same, but we do want to make sure the tests pass in both
environments.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/709244
Original-Revision: 70b83a0c246b45dc6e3275848313d514cd441c8f
GitOrigin-RevId: b5f8817203b6779b71438a660916017a042e89d4
Change-Id: I2b00c8cc89a3e1a80a5f48b72884db60569005f7