[roll] Roll fuchsia [fshost] Remove the migrate_root flag
This flag was added for migrating data to Fxfs before mounting it, but
Fxfs' multi-volume support allowed us to copy data into the filesystem
before we bound it to the outgoing directory in fshost, allowing us to
get rid of this flag.

I also took the liberty of adding an f2fs variant of the fshost
integration tests, since the f2fs migration path was refactored in this
change.

Original-Bug: 93066
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/709744
Original-Revision: 2e670afa2326ba9d2363f9ff9273babe03c7174d
GitOrigin-RevId: bb3e9b606fe8d8651c4195c82f0aac334c262194
Change-Id: Ib5fdbd9e5e753f5414b01b14f3745208daef3cff