[roll] Roll fuchsia [virtualization] Add ZirconGuest to gpu tests.
This enables a simple test to detect displays for the zircon guest. To
support zircon, we also modify the behavior of the rust virtio-gpu
device to wait for the backing view to be created before processing
queues. This is needed because the zircon driver does not yet implement
the config-changed interrupt and will be unable to detect display
hot-plug or resize.

Original-Bug: 102870, 64348
Test: fx test virtualization-gpu-tests --count 100
Multiply: virtualization-gpu-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/708743
Original-Revision: ad6fba89156c0712f2e97893f316c9c08481a2c2
GitOrigin-RevId: 2d7b11a260c2e4e1e6ddc818f239d06dad4f0559
Change-Id: I466364db685cd6535f06f8ae88b35c5f6233010a