[roll] Roll fuchsia [fidl][hlcpp] unknown interactions client.
Update HLCPP client to be able to handle unknown interaction responses.
My approach to doing this is just to expose the result union as a
client-usable type, so this actually requires fairly minimal changes,
and most of this change is updating the HLCPP unit tests to cover the
entire suite of client tests copied from the tests for the new C++
bindings.

Original-Bug: 88366

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/705107
Original-Revision: 6c76eb5d360bc29620f0d17817f9ca3fab189b41
GitOrigin-RevId: 819be318477794f3dab97138458c017b60000e34
Change-Id: Ifec94e73c6438d7528f5d05c787c2cd4247e3f67