[roll] Roll fuchsia [scsi] Check write_cache_enabled bit before flush
Issuing a SYNCHRONIZE command to a disk that doesn't enable a write
cache leads to an error. Therefore, we have to verify that the write
cache is enabled before flushing. In a general solution, we'd have to
check the bit before every single flush, but since we don't support MODE
SELECT commands to change the bit, we can just read it on disk
initialization and cache the value.

Original-Fixed: 86850
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/710162
Original-Revision: 6b3199f6cbd8af029b06fe496962c96c77390477
GitOrigin-RevId: 7c22421dc56f5f7511a6c6a96e84f5e3120ed95a
Change-Id: I3067f288fac87891ad68c106480a46e112b89e4e