[roll] Roll fuchsia [bt][audio-output] Add initial inspect
Add inspect support to SoftPcmOutput.
Refactor so that the audio stream processing is started
automatically when the stream begins to be polled, and the task is
dropped instead of detached.

This also means that if AudioCore disconnects from the RingBuffer or
StreamConfig, the audio stream will stop accordingly.

Original-Bug: 73682
Test: inspect changes only + refactor, confirm with snapshot
Test: fx test fuchsia-audio-device-output-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/664685
Original-Revision: 714caba05e1a8a77b3fcf6dea4ce6b9b0ba19447
GitOrigin-RevId: 4b7c24e56578d733af9dade332a8ea6763fa70f1
Change-Id: I0fb6b872e7f4c4292f5e5b6337abe24cfcc5bcde