[roll] Roll fuchsia [fidl][cpp] Publicize DecodedValue
DecodedValue is a utility wrapper to close handles inside a decoded
object tree. This will be used in the next CLs as part of a standalone
API to encode and decode FIDL objects (FIDL-at-rest).

This CL moves it out of the "internal" namespace and adds a bit of API
polish. For example, |value| is typically understood to return a
reference type.

Original-Bug: 82681

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/709462
Original-Revision: beaf248496e407969065dac229a7008b36b63ac2
GitOrigin-RevId: f74b090b5cf5659e520bcd382700e46bc8986837
Change-Id: Ie3f7d5d2ac3197694b8ec98bab4c2f16f72ac377