[roll] Roll fuchsia [fdf] Remove tools-drivers and tools-system
Remove  "//bundles:tools-system" and "//bundles:tools-drivers".

Originally, tools-drivers was to replace tools-system in order to avoid
driver packages being inserted in the system packaget set. However, it
turns out that the only target in tools-system, a bluetooth hci emulator
driver package, is no longer needed to be loaded into Fuchsia and so the
tools-system can just be removed and not replaced.

Original-Bug: 69738
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/699942
Original-Revision: 2820a519fe16144d2ef16c301912aa527260e355
GitOrigin-RevId: 9f1c00a70a2c91a5303a463b458a31ba3b4eb60c
Change-Id: I54dcd238048908b06469bdb71a8990ecf238c985