[roll] Roll fuchsia [kernel][vm] Do not dirty forked zero page on commit
When a zero page marker is replaced with a forked zero page in a
pager-backed VMO, only mark it dirty if we were looking up the page with
mark_dirty set to DirtyAllPagesOnWrite. This ensures we do not dirty
forked zero pages during a commit, which passes the VMM_PF_FLAG_WRITE so
that it can fork pages if required, but does not indicate
DirtyAllPagesOnWrite.

Test: core-pager-writeback
Original-Bug: 63989
Run-All-Tests: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/711243
Original-Revision: 2a44828d2bda6d8bf76b69e412433a1423e1e1be
GitOrigin-RevId: 8020a53ebee5d4c5185d3f83a7cb9b787028cdec
Change-Id: I5ea4e8cd96799f4ffc6d24906cde7ebbf68373a1