[roll] Roll fuchsia [iwlwifi] Synchronize usage of mvmvif in rx ISR
Unsynchronized usage of mvm->mvmvif[0] in the receive ISR in rxmq.c
led to driver crashes because mvm->mvmvif[0] could be deleted or in an
unusable state due to writes to the object from a separate thread.

This commit adds RCU synchronization and checks to ensure that
mvm->mvmvif[0] is in a valid state before using it in the receive ISR.

Original-Bug: fxb/97275, fxb/98324
Test: Create and delete wlan ifaces 50 times using wlan-dev with
intervals of 0.2,  1, 5, and 15 seconds between creation/deletion,
and observe no driver or firmware crashes. This test was repeated
twice: once without associating with an AP, and once while a YouTube
video was playing in the background after associating with an AP.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/710223
Original-Revision: 597163ccbb33b433010e2f5ef4819c9e574f18d4
GitOrigin-RevId: 32a07a6a05b0c0dd312173f8b4418f010e19fc1a
Change-Id: I806d0429906357f55aa59c80871ecccafd8c5f66