[roll] Roll fuchsia [component][fuzzing] Add 'debug' option
This CL adds a "debug" option to indicate to `Runner`s that the user
wants to attach zxdb or a similar debugger. `Runner`s that do not use a
debugger exeception channel can ignore this, but those that do (e.g.
libFuzzer) need to disable error monitoring when this option is present.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/710828
Original-Revision: 30a7ec12292423370abcd7b99859e55c667fa753
GitOrigin-RevId: a5450b79cbdbc67dd8e07a917342ef9049ed1ad1
Change-Id: Ifc430ae31488bdc00a3fa15948f7a199775b81aa