[roll] Roll experiences [e2e] patch the ermine terminal e2e test for CFv2 monikers
This is a simple fix so the e2e test does not break because it is using
CFv1 monikers. The CFv2 monikers for ermine elements have a randomized
component name, so the complete path is not available to the test, but a
wildcard pattern should work, as the test is currently implemented.

(We are investigating alternative approaches for a future patch.)

Original-Bug: 106471

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/711334
Original-Revision: 48553482c28e223040e8fe2d53a718bbdd4b1022
GitOrigin-RevId: a41ad30f31f8089a3d066cce8a39b5c56b8625df
Change-Id: Ia1498003567ed522bd9a8467923b7f586ebdd685