[roll] Roll fuchsia [pigweed] Configure pw_log backend for DFv1
Create a pw_log backend that uses DFv1.

Make DFv1 the default pw_log backend. The only place pw_log
is needed right now is in a driver.

Reorganize Pigweed backends so that that are under a directory
with the name of the module that contains the frontends.

Original-Bug: 105192

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/704150
Original-Revision: afa60e6d22397caf852649ab9b11e75c832a4201
GitOrigin-RevId: 21f2813406da89d278f6125e9a6efae767a5671c
Change-Id: I2f7d430a9182c858b6c8dea227fd2f69b83a71df