[roll] Roll fuchsia [fshost] Add pause and resume to block watcher
This is used by several integration tests to set up block devices for
fshost without fshost latching onto them before we are done, so it's
useful to get in early.

This also disables a flakey test. The issue is resolved via a more
complex refactor in fxrev.dev/708822 but that isn't going to go in
immediately and it has a high flake rate that is preventing this cl from
going in (and causing other issues).

Original-Bug: 87511
Original-Bug: 106137
Test: fx test fshost_rust_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/701683
Original-Revision: c3fc920471bb2a94b34ff684f9d8fe3b4faa2c46
GitOrigin-RevId: 6d53897c481e6614f1f7931fc056734246557251
Change-Id: I7d0aa4c462d586ae1d40813a4ac323ea2fc38940