[roll] Roll fuchsia [check-licenses] Update metrics template for files
Every package in check-licenses implements the metrics template.

At the end of the run, these template methods are used to collect all
the metrics and results of the run and save them to disk.

With the new feature request of fxbug.dev/104400, I now need to save
the full license segments to disk at the end of the run.

This CL adds a "files" entry to the template instances for all the
packages in check-licenses. Future CLs will actually use this new field.

Original-Bug: 104400

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/707990
Original-Revision: e50f0d27f1d57e9f54d3d25827a1cc056d7a5af3
GitOrigin-RevId: 331ec6c6325791800d48aeddb84cf27c0780d256
Change-Id: I53cb9385405c74c71022e0522951e0f91d4a8ba8