[roll] Roll fuchsia [fdf] Add matched_composite into MatchedDeviceGrp
Adds the matching composite info into the matched device group when
returning it to the client. Store the transform and the matched
composite info in a hashmap that maps the device group's topological
path to the two.

Original-Bug: 105313
Tests: Added unit test for this to the driver index tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/707322
Original-Revision: 67ae8f2dd136fd74a0e9a1bd3a3a679480ef4dfc
GitOrigin-RevId: 7addef3926c8787ad25c19d22527c1a0f0cfb5bb
Change-Id: Icf33d70907d5e9a8f8de41d00a5d7b5f24dc460f