[roll] Roll fuchsia [ffx][daemon] Graceful shutdown of main loop
Currently the daemon exits after a 20ms timeout from getting a quit
message, and it does so by simply calling exit(), which is likely one of
the reasons it winds up leaving files behind in inconsistent or weird
states.

This change makes it so that the quit message triggers an mpsc message
back up the loop to cause it to end normally, allowing the caller (and
any other threads running) to clean up after itself. Also removes an
unnecessary panic! in the event of an error processing a message, also
using the quit message to cause the process to exit.

Original-Bug: 76362
Original-Bug: 106578
Original-Bug: 82518

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/711544
Original-Revision: 59fbdc21530755c92cf588f28f61eb96cac058a5
GitOrigin-RevId: e886dc5c3040c495cd3d15316c997065f6c7c37f
Change-Id: I0927ae77707d7b338fdd0facba3c208a89fb4ff7