[roll] Roll llvm-project/libc [libc] add sprintf size comparison
To accurately measure the size of sprintf in a finished binary, the
easiest method is to simply build a binary with and without sprintf.
This patch adds an integration test that can be built with and without
sprintf, as well as targets to build it.

Reviewed By: sivachandra

Differential Revision: https://reviews.llvm.org/D131735

GitOrigin-RevId: c778ba2795cc6cff7b467c0419924b7333adfc58
Original-Revision: 6a1e429fd8659cc7a4a6e97a01e43f9f8b35a7df
Change-Id: I637a765b2442598de688b61133c2ddc94e0892a0