[roll] Roll fuchsia [a11y] ScreenReaderMessageGenerator: use "out parameters" in Describe helpers
Context: there are a bunch of helper functions in
`ScreenReaderMessageGenerator` that are called as part of `DescribeNode`
to stitch together a description of a semantic node.

This CL changes all such helper functions that had previously returned a
vector of utterances; they now append to an "out parameter" vector
instead.

This is purely a mechanical change. There's no logic changes at all.

This should reduce boilerplate and make it easier to refactor the logic
in this file more in the future.

Original-Bug: 81707
Test: fx test a11y-tests a11y-lib-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/709617
Original-Revision: 54be76b72f63976fe1f8c3dd1c6588706bd2e18a
GitOrigin-RevId: bf86f4c04dc65b8374d2191bfadb38a7c6917e7d
Change-Id: Idd8071cc5e7bf60370ba9dfcd48ff81223213cfb