[roll] Roll fuchsia [check-licenses] Add license texts and patterns to Files map.
This CL adds relative paths to the File and Pattern structs,
so they can be added to the metrics Files map using unique keys.

In a future CL, the "result" package will be updated to write the
contents of the Files map to disk at the end of the run.

Original-Bug: 104400
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/709242
Original-Revision: 593f1907ae0239b5941d0cabaa121ba5fd27bd7f
GitOrigin-RevId: f6d46de5b66e5c38c9c1f7eab3c1195a7128590a
Change-Id: I48cffdca0a92d09e01a08805066fbc6ed616eb02