[roll] Roll fuchsia [scrutiny] Extract structured configuration to a file in the build.
Can be used to regularly check for the addition of new configuration
keys in a system image.

Adds a new integration test which transitively depends on the system
image assembly, so we wire it up using the :host_tests pattern we're
using for assembly tests whose deps can't be included in the graph
of the system image.

Original-Bug: 105366
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/705598
Original-Revision: 3052a05179fdac4d5220dc1b0741c00a1c497e23
GitOrigin-RevId: fa8b54aebd404fcba29a5555e7de8d636bc2835c
Change-Id: I092f8c5671566437f902db6db8d6ab36083ae5cf