[roll] Roll fuchsia [component][fuzzing] Replace ReadFromCorpus with GetCorpus
Early on a decision was made to return inputs from a runner's corpora
one at a time. The exact reason for this decision has been lost to time,
but it led to corpus retrieval being somewhat brittle, and, in the case
of libFuzzer, led to a bug where the first input could not be retrieved.

In practice, callers of `ReadFromCorpus` only ever wanted the whole
corpus, so this CL replaces that method, which took an offset and
returned a single `Input`, with `GetCorpus`, which returns a vector of
all non-empty inputs.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/711504
Original-Revision: 00f867741064a20f07516751b2323ba3549272ed
GitOrigin-RevId: 044c88d4db462743a406c2dd0ce8eb73dcf67bcd
Change-Id: Ied3257f28a7c3874326b56bbf47d8a32b3698100