[roll] Roll fuchsia [fidl][cpp] OutgoingToIncomingMessage requires non-transactional messages
In the next CL we will introduce a non-transactional version of
|IncomingMessage|, a class managing the body portions of FIDL
transactional messages. Non-transactional uses of |IncomingMessage|
would move to that class. Because of that, we have
to make a choice whether the |OutgoingToIncomingMessage| utility
returns the transactional message class or the non-transactional one.

Because more production uses of that utility are non-transactional, this
CL makes that a requirement.

Original-Bug: 82681

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/709463
Original-Revision: 49ab2c992a68b05eb08d5960b057799dd8a1f774
GitOrigin-RevId: d99aff7a3dfb9ae7b8be3a34c58c6a96151d7c60
Change-Id: Iadd5ff3e0ef9ae3d0e509017ec7434885a7e716d