[roll] Roll fuchsia [flatland] Remove unnecessary async dispatches
FrameScheduler::GetFuturePresentationTimes is no longer required
to be called from a flatland thread. That was a consequence of the
old Present API.
This CL removes the callback and as removes a couple of
now redundant async dispatches connected to it.

Multiply: flatland_unittests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/711123
Original-Revision: ede8de92c3b10d9b15f43aec124126f0676924a3
GitOrigin-RevId: 40948ee5a5698bbcfdf25a4a0acdc74a799f43de
Change-Id: Ib1e976b92b4078126e46f911ab0d356f21b88d22