[roll] Roll fuchsia [graphics][compute][radix_sort] Use pigeonhole principle to exit early.
There is an opportunity to exit early in the histogram_rank()
calculation if it's determined that all partitions have a size of one
early in the unrolled partition computation.

Testing reveals that enabling early exit *reduces* performance on
devices that use the unrolled ballot-variant of the partition
operation.

For this reason, the switch isn't enabled for any device at this time
but may be in the future depending on the device and codegen.

Other changes: Some of the nearby comments are shifted downward
because `clang-format` improperly indents comments if they're above an

Testability: No change in behavior but the switch was enabled and
tested locally on discrete GPUs.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/711807
Original-Revision: c528bc8ddb7163dc814e23b8a5186e04902c7af1
GitOrigin-RevId: 10ab15e1d028787cc2f72cf760231a8d8f88bb22
Change-Id: I24a3b47f00978cad4b0624a4f35d586762044149