[roll] Roll infra/recipes [sdk] Add property to set Jiri attributes
This is necessary to support conditionally including arm64 emulator
binaries in the checkout. We want to include them in SDK builder
checkouts so the binaries can be included in the SDK, but not in default
checkouts since the binaries are quite large and are updated often.

Original-Bug: 99321
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/712407
Original-Revision: 200847e89e0ae66bd711cf574b01ab54ccd721ef
GitOrigin-RevId: e6733c11e4fc21bb88b7a5dec3321ef5aa3b1bd9
Change-Id: Icccc853c1abea4f7a6b778cc1dc0224bb049afbf