[roll] Roll fuchsia [zither][go] Introduce go data layout bindings
Partially as a demonstration of the utility of the machinery we've
built, we add a go backend to zither with minimal effort. It will serve
to have something not entirely C-like as a reference in further
development, and this backend will later be needed to deprecate kazoo's
generation of bindings for the go runtime.

Original-Bug: 91102
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/708763
Original-Revision: 3aad47410e0abc91f9d89c5494026cecb410cbaa
GitOrigin-RevId: 6fbe3a6a751f2933d927023db0a362813c98d952
Change-Id: I220d8005b7ae5268a1e47436b673f9f69d465983