[roll] Roll llvm-project/libc [libc] add guard for file pieces of printf
In the printf_core CMake, the file pieces are defined as object
libraries that depend on the File data structure. If these are added
unconditionally they'll try to evaluate that dependancy even when there
is no File available. This patch adds a guard to prevent that error.

Reviewed By: sivachandra

Differential Revision: https://reviews.llvm.org/D131921

GitOrigin-RevId: 3449b7cd59401b1e5bf0590b51bf6109ce049c64
Original-Revision: 3f36b411459e152cadb1cda4997bb3dd51f8e6d3
Change-Id: Id54101e40bf3de87896e8b6eef5128c01fd387bd