[roll] Roll fuchsia [pbms] Generate product bundle for bringup
This CL enables the generation of a product bundle when building
the bringup product. The schema is adjusted to allow for an empty image
list, since bringup does not generate an FVM, and the `ffx emu` tool is
likewise adjusted to accept an empty FVM list.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/712545
Original-Revision: 639442e11dd3bb64005012cd9ea4ab5222b619f7
GitOrigin-RevId: 88ef01aeaa61f2e48a83c2d82b8033e5cfe3df58
Change-Id: I9550c0e974d1c2f7c1242108a697bb015eb3707e