[roll] Roll fuchsia [build][test] Fix package_name for fuchsia_test
This CL corrects a discrepancy between the `fuchsia_test_package` and
`fuchsia_test` GN templates when a `package_name` is provided. In the
former, the `target_name` is used to derive the `package_manifest`
before the check to see if `invoker.package_name` is defined. The latter
constructed `package_manifest` after this check and thus used the value
of `invoker.package_name` if defined. This mismatch caused tooling
designed to work with `fuchsia_test_package`, like `test_list_tool`, to
fail if `fuchsia_test` was given a `package_name`.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/712806
Original-Revision: e063e478922379da0aa94c2df8996705b2a7af9a
GitOrigin-RevId: fb9759a58fa0a361c390ea46b8d27a6119f65395
Change-Id: I1421dc584935c508f71b134d08494fa4f0c6cd21