[roll] Roll fuchsia [fidl][cpp] Natural types printing
This adds the ability to print natural objects. Because this could add
bloat to programs that don't need it you need to add an explicit build
dependency on libraryname_natural_ostream and include the
fidl/libraryname/cpp/natural_ostream.h header.

The printed format is not JSON or C++ or FIDL. It is not stable and will
change. It must not be used for tests.

Documentation will be coming in a follow-up CL.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/702513
Original-Revision: fa984b1219f4bd400cce72d75017eeff754da391
GitOrigin-RevId: b3077658eb1f42b1e9aab6321d42a7c1cdaca687
Change-Id: I6ccce64293581076a04eb9fc65aec2eed7c6b8c4