[roll] Roll fuchsia [kernel][irq] Check IRQ config when DEFAULT config is passed.
If a user attempts to create an interrupt and specifies "default" for
the trigger mode and level of the interrupt, be sure to check the
current configuration of that interrupt before configuring the
interrupt dispatcher.

Failure to do this means that we might have an interrupt which has
been configured for level triggering (because of ACPI), but then fail
to note that we need to mask and unmask the interrupt at the start/end
of processing.

Original-Bug: 102058
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/712703
Original-Revision: cddb5d12a3d16d33b00f14e2855a42282e76530d
GitOrigin-RevId: 4df401d6d7018af169500e663d6c00f837ed6b91
Change-Id: I22034128a64a2f55feb66899d79d1000a38eeb81