[roll] Roll fuchsia [kernel][irq] Fix a bad range check in remap_interrupt
remap_interrupt on the PC platform was checking to see if the
interrupt vector being passed was in the range from [0, NUM_ISA_IRQS],
and passing the vector on to apic_io_isa_to_global if it was.  In
several places in ioapic.cc, however, it is made clear that the proper
range should be [0, NUM_ISA_IRQS).  There are several places where
the code ASSERTs that vector numbers are strictly < NUM_ISA_IRQS, and
the vector itself is frequently used as an index into statically sized
arrays of NUM_ISA_IRQS in length.

So, fix the check in remap_interrupt so that the range is correct and
does not accidentally trigger an assert.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/712704
Original-Revision: a58f56b2824341700a66f37ab1e6f559de17209b
GitOrigin-RevId: dfe4455b908db746dbdb643c6dfd87e49d5df394
Change-Id: If2010a959c9eaad80885724470479e9c5c4add62