[roll] Roll fuchsia [fit] Fix internal class template parameters
The function_impl, callable_impl, and function_base classes were
declared as if they had Result and Args... as template parameters, but
this was not the case. These classes are actually templated on the
callable type, and the result and argument types are extracted in the
partial specialization for function types. The Result parameter was
used for the callable type and the Args... parameter pack was always
empty.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/710573
Original-Revision: fa85e608bba04b8e7e316f8159b29fa0b735f6ff
GitOrigin-RevId: c4dc12ed2ca685b61839c38b386b7dfd91181a77
Change-Id: I470e7a11d37ab676f7c9f2f050026f66aae3f86f