[roll] Roll third_party/go [fuchsia] lookup profile gp from sp
Relying on only mp.curg to find the correct G for an M is inaccurate and
incomplete. Duplicate the logic from os_windows.go to determine the G
from SP to cover all of the cases. Note that Windows uses a similar
profiling mechanism (profiling from a remote thread) to Fuchsia.

Note that if a thread created in C calls a Go function, it receives an
"extra" M for the duration of the call, which is
initialized/deinitialized with minit/unminit. We don't want to profile
these Ms when they are unused because (a) the contents of the M are
meaningless and (b) we will fail to find gp from the SP, which may crash
sigprof. Thus we must clear mp.thread in unminit. profileloop already
skips Ms with an invalid thread handle. Again see corresponding
os_windows.go logic.

We have to add locking around unminit to ensure that profilem can't race
and see a thread after it completes unminit, at which point the SP could
be anywhere in C code.

This seems to move the needle forward for fxbug.dev/72646 and
golang/go#40823, but we're still observing flakes there. This change
unblocks bringing in go 1.18 by removing usage of `sigprofNonGoPC`.

Multiply: netstack-bench-gotests :45
Multiply: go-pprof-test :45
Original-Fixed: 90865
Original-Bug: 72646, 93043
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/third_party/go/+/629462
Original-Revision: e058f3040db55dac169c2714bc3a88ebbef05371
GitOrigin-RevId: 3f2e2ae1fe3bf213adb8738d12e8c15ff1b60617
Change-Id: Ie3293c73168056433c5adc7a362fdc22c5da91ad