[roll] Roll fuchsia [settings] Match background stream via settings
Sets the background audio stream to be at the same volume level as the
media or interruption stream when either of the two are modified. This
is in preparation for moving this logic over from another component.

Original-Bug: 104024, b/240315145
Test: fx test setui_service_tests,
volume_change_earcons_integration_test

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/711335
Original-Revision: f5fac5746a8d008dc1459216bc24dd27eced4235
GitOrigin-RevId: 6d781b27da3e536736708720f8bfd07951590199
Change-Id: Ibe59ebdffd099b8f3885930de261708b76a9a057