[roll] Roll fuchsia [ntr] Add support for starting DHCPv6 client
Add support for starting DHCPv6 client when requested. Only one client
can be started on an interface at any given time. This limitation
shouldn't be an issue, as the RFC recommends against running multiple
clients on an interface, and so it's unlikely this will need to be
supported in the future.

Changed DHCPv6 client to block `WatchAddress` forever rather than
closing the channel and stopping itself so that it's possible to
exercise the watch stream.

Add ffx plugin test and an integration test which verifies that the
initial Solicit/InformationRequest message is sent by the DHCPv6 client
when started in stateful/stateless modes respectively.

Test: fx test --test-filter 'start_dhcpv6_client::*' \
              network-test-realm-integration-test

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/707509
Original-Revision: 4655478c4e55a987e5fdeddc769b0f940c6a686d
GitOrigin-RevId: 2716a200ea738d810bce871c34ed698ff9aba7e3
Change-Id: Ie12f1acbda7bd427ecf1168cab922eccaa54781d