[roll] Roll fuchsia [media/playback] loosen rigid ordering in tests to eliminate flakes
Tests that use FakeSysmem submit expectations for the buffer
collections that will be created and the constraints that will be
submitted by their respective constituents. The expectations
submitted to FakeSysmem consists of a list of
FakeSysmem::Expectation objects, each of which describes a buffer
collection that is expected.

Previous to this CL, FakeSysmem assumed that BufferCollections would
be created in a specific order, namely the order implied by the list
of expectations. While the order is *usually* predictable, the order
is not guaranteed, and tests would fail from time to time.

Rather than requiring that buffer collections be created in the
expected order, the updated tests require that each buffer collection
match exactly one of the expectations in the list submitted to
FakeSysmem. This is a somewhat weaker test, but is more than adequate
to detect regressions.

The tests were run using expectations listed in reverse order to
ensure that the ordering requirement has been eliminated.

Test: change to tests - no behavior change

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/713263
Original-Revision: bdcb3033744d0cdb31c0f02a0f9e7b783f065262
GitOrigin-RevId: bbd0a0e3b4336b832d278f45f7577cf379e589fd
Change-Id: I8aae39bd9b6f017672675be05a1b7aef2f6b4104