[roll] Roll fuchsia [emu][sdk] Adding aemu arm64 prebuilts to SDK
This adds the aemu_internal for arm64 atoms to the core SDK.

The prebuilts have been added to the manifest and roller updated in a CL
in the infra repo.

They are optional for a developer checkout, so to enable generating the
arm64 files in the SDK,
Run: `jiri init -fetch-optional=arm-sdk-tools` and `jiri update`.

Then include `--args arm-sdk-tools=true` in your `fx set`.

qemu arm64 is pending on fxbug.dev/106608, so will be added in a
separate CL.

Original-Bug: 99321
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/711264
Original-Revision: ccd39b5ffb11c75b809c91f03c2a5807b8bcc986
GitOrigin-RevId: c5caf5c79aa4b285598a68a308f8e2dab102ab9d
Change-Id: I5db40f0cc891e92fc7c09803e0cb7e7343870ec8