[roll] Roll fuchsia [bt][core] Remove peer from denylist if we try to connect
When a peer is trying to connect to us after we disconnected due to
local choice, the denylist avoids overzealous autoreconnections from
succeeding.

Unfortunately sometimes we want to connect back to that peer, but they
will not accept our connection and insist on connecting to us.  When
that happens the peer will reject our outgoing connection and continue
to attempt to connect, and we will be stuck in a loop with no way to
connect.

When we attempt to connect via local choice, remove the peer from the
denylist so they can initiate an incoming connection.

Original-Fixed: 102450
Original-Bug: b/228972015
Test: updated unit tests; fx test bt-host-gap-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/713522
Original-Revision: e2013bdfba18f86a95178d3643e3d795f3f7f9e1
GitOrigin-RevId: dba01deb8ef3028546bed10165be28a05d2bbdc7
Change-Id: Iae260417c5a9b655e5d0d4efa660d207d278d1be