[roll] Roll fuchsia [benchmarks] Remove the disabled touch_input_latency test
This was added in 2020 but was disabled not long after it was added.

It is one of the few remaining performance tests that refers to a
Components v1 component.  It refers to touch-input-test.cmx, which was
removed (and converted to touch-input-test.cml) in
https://fuchsia-review.googlesource.com/c/fuchsia/+/559102.

Original-Bug: 86575
Original-Fixed: 76494
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/709816
Original-Revision: 8806217bf8e69571a711da984dc7922d4eb8cd97
GitOrigin-RevId: de2fb78022eafe18de82f435501085bbad594fd3
Change-Id: Ib04f25141fe248eb59f3efd51ec98bc5ce5a5e80