[roll] Roll fuchsia [overnet] Add new torture test to QUIC
This also simplifies some of the locking logic in a way that shouldn't
affect functionality but made the code easier to read about while
debugging.

It also handles an error condition when we bind a remote stream and
immediately attempt to write to it even though the other end hasn't
established the stream yet. The latter shouldn't affect much either;
producing a situation where this happens and behavior is correct is much
harder than I initially realized, and I ended up deleting the test case
that would have exercised it because it was not and could not be
correct, but the change and the comment are worth keeping for posterity.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/710662
Original-Revision: d1c4f797783c5c2570594d2c0b60c19f71b37f72
GitOrigin-RevId: 97e87d61d432d9298092091f3fb52a765be78232
Change-Id: Ib81a5ed84f5f496801933c999f0da5e1a6b9ca30