[roll] Roll fuchsia [sysmem] Extract common FIDL messages to Node
With this change, the semantically common messages will
automatically get migrated to the "compose Node" messages
next time a sysmem client is rebuilt.  Clients that aren't
rebuilt yet will still use the old messages which are still
implemented (for now) but which now have names that make
their deprecation obvious.

Going forward, new messages that should exist for all
protocols that "compose Node" will be added to Node.

The implementations of the old and new variants of the same
logical messages are shared to ensure identical behavior.

This is in preparation for adding another protocol that will
"compose Node".

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/710569
Original-Revision: 7153cd84bedc69eb92e91cde74e38e02ba83e87b
GitOrigin-RevId: 05d08d13a33c073924905a680bde0cdb9d9a2642
Change-Id: I29dcad99d9d9a2c085c8f02612fb2b5a08a8a9d7