[roll] Roll fuchsia [testsharder] Don't skip tests if there are no affected tests.
A change for which we are unable to determine affected tests would tell
testsharder there are no affected tests and it would end up skipping all
hermetic tests as a result. Instead, we should assume that if we are at
the testsharder step, we've determined that we want to run tests (either
because there are affected tests or we want to run all tests), so if we
see there are no affected tests, we should just run all tests.

Original-Bug: 107088
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/712384
Original-Revision: 41d00b31e58fbbcd0ee99f6a9df2ba3f518812c3
GitOrigin-RevId: 76e24e52246dfe9dbf97ea40896aee676ae8bfd3
Change-Id: Ibff6c9840aba8758c97f75be767f014cea7da75e