[roll] Roll fuchsia [syslog] Do not inline LogNew
We can encounter a stack overflow in blobfs that prevents networking
from working if hwasan is enabled. The codegen caused 4 LogBuffers to
be added to the stack which would decrement SP by ~131kB. This
behavior can also be seen with asan without the stack-to-heap
promotion. This occurs from a combination of inlining, a large
LogBuffer, and sanitizers not allowing stack coloring to work
correctly. Until we can get hwasan/asan to work with stack coloring,
or the syslog API can be reworked to avoid adding a large buffer on
the stack, we can unblock the hwasan rollout by marking the function
that creates the LogBuffers as inline. This may likely be a small
penalty since this is usually done in a failure handling path.

Original-Bug: 106574
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/713441
Original-Revision: e9cff66016d1a3458eed18e0495aa257ebe10246
GitOrigin-RevId: 8284670408c315f713a3c574ea438860b262ee16
Change-Id: Id035724856e1be54896c6bfba47b4d9250a66f01