[roll] Roll fuchsia [usb][xhci] Use deadline profiles for XHCI IRQ threads.
The XHCI threads responsible for handling USB interrupts (the
"completer" threads) were using high weight fair profiles, instead of
deadline profiles as they should have been.

Using a high weight fair thread is like saying that you need a lot of
throughput, but that latency is not so important to you.  A deadline
thread is what is needed when latency is the main issue.

Also, instead of hard coding the profile definition in the source
itself, move it core profile definitions and then apply it to our
thread using the profile name.  This will not only make it much easier
tune the profile if needed, but also to override it on a product by
product basis if needed.

Finally, do something similar for the USB audio callback thread.  It
was using a proper deadline profile, but the profile was hardcoded.
Now, it is defined by central config instead.

Original-Bug: 102058
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/708523
Original-Revision: 6719aec2ff3dc7851de900a991b0d4b205684799
GitOrigin-RevId: 6e4f3c55895db4f03930b27bd4721883e427c056
Change-Id: I47f85adc94e9ec1d78ccc8c78787b5aa721f45d9