[roll] Roll fuchsia [system-updater] Fetch images from images.json if present
If an images.json manifest is included in the update,
fetch the images that need to be written. Be sure to
use the resource provided to find the image rather than the
current assumption of imagetype.name().

Update the retained index at each step to guarantee forward
progress.

Adds tests for parity with the soon-to-be-deprecated
update package format.

Original-Fixed: 95804

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/704011
Original-Revision: 8ca1c66f9bff3c6ceea13d924719a905878a9bb5
GitOrigin-RevId: e1127492a62f78892a4a72e751e8be06cdc4f888
Change-Id: I1adadc76f83e74fa92e198464e1dad58e09c513f