[roll] Roll fuchsia [wlan] Pack BitField types
Clang has changed it's behavior in attempt to match gcc's
by not packing non-pod types unless those types are
explicitly packed also. In this case the BitField types
were not marked and are not pod. LittleEndianBitField can
be trivially made pod by making it's data members public.
BitField would need to not default initialize val_ to 0,
which is untenable. So instead the address-of-packed-member
warning is suppressed to mark these types as packed.

Marking these types packed will keep the same layout across
Clang abi versions. It was a bug in clang’s
address-of-packed-member warning that it wasn’t warning
before. ASan caught this UB which is why -fno-sanitize=alignment
is currently being used, this is a known problem in the code
so suppressing the warning is sensible.

Original-Bug: 92521
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/713825
Original-Revision: 3a2fde80f0de513d6f2accd646cc2c3aa2d19183
GitOrigin-RevId: e65aed9074b7d8d393e87075102d3a0cbbf016d3
Change-Id: Ia8130fc4e97222a76c643e0acecae339d9c80379