[roll] Roll fuchsia [kernel][wait] Fix a wait queue ordering bug.
Fix a bug in the selection routine used for wait queues.  Basically,
there was a situation where if you had a wait queue with at least one
deadline thread in it, and at least one fair thread, and all of the
absolute deadlines of all of the threads in the queue had expired
relative to the query time, the Peek operation would return a fair
thread instead of the deadline thread with the minimum relative
deadline.

The fix for this is simple.  Also add some tests which check all of
the basic ordering properties of the collection.

Original-Bug: 102058
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/713514
Original-Revision: f8546cd05a5752ceffc1bffa2c563ad82abe19ea
GitOrigin-RevId: f3f7d03b3326ae07bc628c127be7084d0a53e45a
Change-Id: I9b8c5593fe84118c6f7403c66ea123f8785da901