[roll] Roll infra/recipes [btu] Symlink deps into a tempdir before uploading
The `cas archive` command doesn't follow symlinks, so symlinks in a
test's runfiles that linked to another file in the bazel workspace
wouldn't have the link target included in the CAS inputs, resulting in a
dangling symlink when the task ran.

To work around this we need to reconstruct the test's deps in a
temporary directory with all files copied rather than symlinked. This is
pretty inefficient, but we do the same thing for in-tree testing.

Original-Bug: 106189
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/713583
Original-Revision: 8e6c52b3651fb502886250790ff2693737011eb2
GitOrigin-RevId: 1a70e08de20fff342a2bd32be02a6e6bf22242a2
Change-Id: I11dc7daa9acfbab5979ed9e739e883b72193336c