[roll] Roll fuchsia [netstack3] Buffer creation for active open connections
The bindings need to pass down an object (zircon socket on Fuchsia) to
Core to create buffers later because `Describe` could be called before
calling `Connect` so the Bindings must have created the zircon socket
pair before initiating a connection. This change makes the state machine
be aware of the difference and tracks the object. The buffer creation
for the passive open path is unchanged. `ClientEndBuffer` doesn't make
much sense in the active open case, splitted it into two types:
`PassiveOpen` and `ActiveOpen` which are both `()` in this CL.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/712589
Original-Revision: 147ffe357032785af7ba08f1966d353b1021bd95
GitOrigin-RevId: 255d33b1d59cdabec514d14a5f681839eb899694
Change-Id: Id4bf2475c44a382df1e1aee454861071a3ddac4c