[roll] Roll fuchsia [ffx][config] Clean up bad lockfiles proactively.
Do a few simple heuristic checks to deal with the most common situations
where a lockfile might be left lying around and delete them when we can.

This is both somewhat conservative -- it will only delete for specific
situations -- and somewhat aggressive -- in that it assumes it's better
to have incorrectly deleted a lockfile that was in active (but slow) use
than to block someone. Since these are actually just config files with
wide validity bands in terms of time, this is probably fine.

Original-Fixed: 107136
Multiply: ffx-e2e

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/712992
Original-Revision: b7a2d27e87535451ccfc330efe771988dc2a6b16
GitOrigin-RevId: 1ec3aa6b261968f681aa337e0f92f02134e063ea
Change-Id: Iadd37fa530d694f867fe247200c729f0f46af504