[roll] Roll fuchsia [fxfs] Remove downcast from Allocator to SimpleAllocator.
This removes FakeAllocator in place of SimpleAllocator and a whole bunch
of dynamic polymorphism that we don't need. If we ever want to support
more than one allocator, we may have to put some of this back but it
seems unlikely.

Original-Bug: 96077

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/713849
Original-Revision: b0d362e5fa28f7186cc5396fe4733101b0f70157
GitOrigin-RevId: 286d6b030c69f479f028c58d09aa94f36426625a
Change-Id: I3d79379637322458f5b1363e2425005f886db01f