[roll] Roll fuchsia [appmgr] Fix namespace for components using scheme-Runners
Components launched with a Runner based on the component's URL
scheme were previously launched in an empty namespace, with access
only to any |additional_services| specified in the component
launch request.

The |additional_services| are now used to augment the Namespace in
which the component runs, and a FlatNamespace constructed which
populates the "/svc" directory from that Namespace's services.

Original-Bug: b/241753768
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/713965
Original-Revision: 04696a48ed515da1537647386a7a7414d106d041
GitOrigin-RevId: 73c11ee79b1d80ed7225fcbf24039c13847314d9
Change-Id: Icc4468cb775dd8eb34b8da6da362bafe3005fd0e