[roll] Roll fuchsia [netstack3] Add IpSock options field
Add a field to IpSock to hold options. This is currently empty, but will
be used to record multicast send options, e.g. the equivalent of
IP_MULTICAST_IF. Making this a template parameter makes it easy to limit
multicast options (and the cost of recording multicast options) to
usages that need them.

Original-Bug: 103426, 39479
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/712368
Original-Revision: 8823aa5238ffd1f64b4843ba494ca93fcc806138
GitOrigin-RevId: 80f3d9c824df3d53d8c9ac5ec2f6b9029d0917e9
Change-Id: I2484c9cb512fae6d4c7973556022c6ddc1734de5