[roll] Roll fuchsia [network/tests] Test async errors on transferred dgram client
Validate that ICMP errors triggered by a socket IO operation can be
received on a new datagram socket client created by calling
`fdio_fd_transfer` on the original fd.

In Fast UDP in particular, sockets use intertwined logic to 1) check for
errors and 2) check the validity of a client-side cache of destination
addresses. This test verifies that even when a socket is sending to its
connected address for the first time, and therefore has no cached
address for which to check validity, it will still correctly propagate
errors to the caller.

Original-Fixed: 103655
Test: fx test -d netstack_with_fast_udp_dgramsocket_c_api_test
      fx test -d netstack_dgramsocket_c_api_test
      fx test --host netstack_dgramsocket_c_api_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/713084
Original-Revision: b296883228d6136e9a53914c28dbecba916079e1
GitOrigin-RevId: 46f4eeec8bf309835d6f38cf097b11cf3acca20a
Change-Id: Ifeccc0ead6fdc0c1df2243ee1aacd004511586f2