[roll] Roll fuchsia [debugger] Communicate target system page size.
Some operations on the client need the target system page size. This
adds that information to the "hello" message along with the architecture
and stores it in the client in the ArchInfo class.

Hooks this value up to the "stack-usage" command (removing the hardcoded
4K constant) and the "aspace" command (now additionally showing
committed bytes).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/714342
Original-Revision: 618bf452bc28219915703c1f7fffb62efceec67a
GitOrigin-RevId: 234f1f970ef1ccf05df1543e694b2b205444414b
Change-Id: I9faa3c076c21eb1f53fd578814a7cf986a605ebe