[roll] Roll fuchsia [fidl][cpp] Separate transactional message from non-transactional message at the type level (1/2)
This CL and the next breaks up IncomingHeaderAndMessage into two
classes:
- IncomingHeaderAndMessage represents messages prefixed with a header
- EncodedMessage represents a message (without a header).

PS1 and PS2 moves IncomingHeaderAndMessage to a dedicated header and
adds an empty EncodedMessage class. The next moves functionality over.

As of this CL, EncodedMessage is embedded within
IncomingHeaderAndMessage, but it still represents both the header and
the body. IncomingHeaderAndMessage just delegates operations to
EncodedMessage. The next CL will change it so that EncodedMessage is
offset forward by 16 bytes.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/709973
Original-Revision: c464c4288a7ac195baba70859577dfb9cbfe1bd0
GitOrigin-RevId: cabd6aa1f5a98d1493bc60b480b85062f89bc7ea
Change-Id: I2af1b1d6476d9fc4683a7bf74dccd1651b6f5544