[roll] Roll fuchsia [wlan][rsn] Adjust SAE retry parameters
The SAE retransmission timeout is raised from O.2 to 1 second,
and the maximum number of retries is lowered from 30 to 3. The
extended timeout will improve compatibility with APs which sometimes
take longer than expected to respond to an SAE Commit message.
The lower number of retries prevents ending up in a long cycle
of sending SAE messages to an AP that clearly will not accept them.

At this time, we don't currently have a straightforward method to test
this timeout in an integration test.

Test: Manual test with previously incompatible AP.
Original-Fixed: b/238696147
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/714102
Original-Revision: 36dec1b4b9b5092aef8ffe3f94c1d2f0f60f6e62
GitOrigin-RevId: 47e62f26533893e324352537ccec977c57b12572
Change-Id: Id7533f22d7ff841cf93954070fa5ae8bd48990ea