[roll] Roll fuchsia [ffx][product] fallback to prior PB loading
Be clearer about what's failing and automatically fallback to
using the old methods of loading the PB if the local path fails.
It will report this attempt and failure to the trace log, but
doesn't tell the user. It silently goes on as if the local file
path were never attempted.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/714709
Original-Revision: 0a4ddee72180399277be9a392e9ab77b5b5f5c86
GitOrigin-RevId: 8a5f8584b492c1a3ce7664fe1eb188176c758f93
Change-Id: I5b64464da8c64a80e195ee1fad187f0c1ae2e16b