[roll] Roll fuchsia [check-licenses] Rename filetree to directory.
The FileTree package just represents a directory / folder structure.
It was named as such because check-licenses uses it to create a tree of
file objects in memory during the traversal, but in reality it
complicates the check-licenses architecture.

This CL renames the "filetree" package to be "directory".

I haven't deleted the old filetree folder yet, as I want to keep this CL
as small as possible for easier review.

Original-Bug: 105890
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/714391
Original-Revision: 5e9d05c19e3af0a5b74f5bcdcb1bc665fae8875b
GitOrigin-RevId: 184ebec7974c01b8bc67f756a609f4ddf06df5d2
Change-Id: I7daa0435fa5627ee777df030d62d91d84d37771b