[roll] Roll fuchsia [ffx][test] Stabilize output step 2: support stable output
This is the second in a series of changes that stabilize the structured
output format produced by ffx test. These changes make breaking changes
to the experiment flags supported by ffx test. These changes are made
in a series of steps to soft transition known users of the ffx test run
command that rely on structured output. Any other uses of ffx test
experimental features may be broken.

In step 2, the meaning of the test.experimental_structured_output flag
is changed. Before this change, the flag needed to be set to access any
form of structured output. After this change, the flag indicates to use
the experimental format schema, and absence of the flag indicates to
use the stable format schema. Developers that set
test.experimental_structured_output=true continue to see the
experimental output, and developers that have not set the flag continue
to not see any output. After this change, known infra users will be
migrated to use the stable output. This also adds documentation on how
to use the new output.

In previous commits:
* Step 1: Decouple experiments that are not ready to be stabilized from
  test.experimental_strucutured_output

In future commits:
* Step 3: Remove the test.experimental_structured_output config flag

Original-Bug: 81195
Testing:
Manually verify that this produces the stable output:
  ffx test run <url> --output-directory <dir>
Manually verify that this produces a deprecation warning and the
experimental output:
  ffx --config test.experimental_structured_output=true test run
    <url> --output-directory <dir>

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/701242
Original-Revision: e8f22f3f9196b0700fbc34a7ac3693a36ab2e13f
GitOrigin-RevId: 23e8c469b69a1efcc01de1fd3bc59b709e8987e8
Change-Id: I2f42b0a5d577c8f0937153d07b34baafd71d0ee2