[roll] Roll fuchsia [benchmarks] Add metric name expectations files for iperf networking tests
This test does a large number of invocations of convertResults(), so
unfortunately we end up with a large number of expectation files.
Fortunately the files are not large.  (The test is instantiated with a
large number of parameter values, so maybe this could be pruned back
in future.)

In the test, repurpose the "label" argument to contain a fragment of
the expectation file name.

Change the expectedMetricNamesFile arg to convertResults() to be a
named argument for searchability and for consistency with other
functions.

Original-Bug: 105202
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/714175
Original-Revision: 780e9b3828da4ddd8b19cf985d02771459f6fe09
GitOrigin-RevId: 644646579a1a65aae267569749b9267c279947b4
Change-Id: I9496e1781c3be7608f24c3d98f19bbbef9601bbe