[roll] Roll fuchsia [lowpan-ot-driver][discovery_proxy]: Work-around for missing services.
This is a work around for what is fundamentally the same bug as
<fxbug.dev/99755>, where the child channel will unexpectedly close
when the channel that was used to create it is closed. The work-around
is to keep that parent channel alive while we are still using the
child.

Original-Bug: 99755
Original-Bug: b/241818894
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/715287
Original-Revision: b98e619afe8ff705628cee9da50007453520de68
GitOrigin-RevId: 8b12c7950a22f5bf7d7087c718716b388057fc51
Change-Id: Id2c12f5ca253ce2bc821f9ef65912df2972868f3