[roll] Roll fuchsia [kernel][x86] Save registers for suspend.
Continuing to add support for the S3 (Suspend to RAM) sleep state.

Part 2/3. This CL saves x86 registers before calling on a function to
transition to the sleep state. That function is not yet supported.

To follow is filling out that function to set the PM1 register with the
sleep bits.

Original-Bug: 107532

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/587301
Original-Revision: c4d49dd1a52af1e9f4511d0ef302113c2606c265
GitOrigin-RevId: ca0186f6c617a6505177d5aad6e6f41582007f04
Change-Id: I52b7f2e8d1b74e85a699a2179a052414c53b310d