[roll] Roll fuchsia [syslog] Remove LOG_SEVERITY_AND_ID_*
While well intended, this feature remains largely unused today.

As part of improving and simplifying the sdk/lib/syslog API before
placing in the SDK, remove this mostly unused feature. We can bring it back if desired in the future with SDK-level design or experiment with it in-tree through a non-SDK library.

Original-Bug: 105520

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/705662
Original-Revision: 539f1fdf815999b94f3503ece0cf33b710390f88
GitOrigin-RevId: c4ab16cb57e890365db0ecadfbdedfbf9bd46d45
Change-Id: I7da85a220726b17ecf900dfafdeefac6a2c75566