[roll] Roll fuchsia [fuchsia.process] Add SetOptions to Launcher
This allows clients to set the options with which the process is to be
created.

This change also includes updates to the elf runner implementors to
support the ZX_PROCESS_SHARED option. The flag is set by adding
`is_shared_process: "true"` in the component's .cml.

This change allows the Starnix runner component to be created using the
ZX_PROCESS_SHARED option. Starnix needs this flag to run components in
"restricted mode."

Original-Bug: 91399, 98922

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/714824
Original-Revision: bb71f23bb1084d0c0a8d7e27fa597a4624852d20
GitOrigin-RevId: 179b530fa5f6a6e0888dac00b649e88b31dcdf31
Change-Id: I34f686f6ff1752c4a3334fd4b462dd3000ecfc49