[roll] Roll fuchsia [check-licenses] Move utility functions to separate package.
The git file handles running the external git binary to retrieve
information about a repository (URL, git commit hash, etc). This is
needed when generating the compliance worksheet, as we need to provide
hyperlinks to the upstream projects. This CL moves that functionality
into a separate package, so it can be used in multiple places across the
check-licenses project.

Same goes for the GN tool and its usage across the project.

This is mostly a copy-paste from the result/world/* directory.

Original-Bug: 104400, 103596, 105890
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/714145
Original-Revision: 266627bfebb634d20635718e02a25d1629635d31
GitOrigin-RevId: 4f0f4169f293b9fe91b9aba1c97a92f39c15bc7e
Change-Id: I1658dd1da19960a67983b5e3028bedf8c404b9c1