[roll] Roll fuchsia [emu] Fix upscript parsing
The previous CL changed how the upscript path was being retrieved from
the ffx_config, and in doing so it broke the workflow if the value in
ffx_config was an empty string.

This CL handles the empty string case as well as the None case, and adds
a test to guard future changes.

Original-Bug: 105286

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/715514
Original-Revision: fac268ace04ec632ca836b92d30de5733a0e149e
GitOrigin-RevId: 467a82ed91b7794eee38eca9a420ad79f61291c0
Change-Id: I58b40f67b2228a754dc95895bb2f0df5aedb32aa