[roll] Roll fuchsia [ffx][self-test] Correct daemon-config-flag test
This makes it so that the following happen in the daemon config-flag
test that were previously not happening, and likely leading to
false-positives of this test (as well as sometimes false negatives):

- Wait to make sure the daemon actually spawns and stays spawned.
- Wait to make sure the daemon doesn't die after the quit message is
  sent to the 'wrong' daemon.
- Don't do a blocking wait in the async block, preventing things from
  moving forward (Child::wait).

This change also makes it so that the output for self-test changes to
show that a test has started and then only prints that status after it's
done. This is just useful to make it so that you know the test really
started and that it's not stuck on cleanup on the previous test or
something.

Multiply: ffx-e2e-without-target
Original-Fixed: 105903
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/715605
Original-Revision: 981c7a26cb9eff66017baa8277f5a0e9976e2e92
GitOrigin-RevId: 22a1053f05c449fb52ce4953c4f7b9332f916d1d
Change-Id: I1e5de6a10a07b7155413f1e65c789b1f92199855