[roll] Roll fuchsia [hypervisor][x86] Change how we perform invalidations
Firstly, make sure that we invalidate an EPT across all CPUs when we
first create it, to ensure that if we are reusing the address of a
previous EPT that all cached data is invalidated.

Secondly, don't invalidate EPT (host physical to guest physical)
translations when we migrate CPUs or when we create the VCPU, only
invalidate VPID translations.  The EPT only needs to be invalidated when
we modify it, but we may have stale VPID translations cached on the
target CPU.

Multiply: hypervisor_tests
Multiply: virtualization_core_tests
Original-Bug: 95763
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/715462
Original-Revision: 7d5312bdd242e3301d98150baffe8d281331c04d
GitOrigin-RevId: 553a3f6599fbc3c99aa7c0f823242f3a5abc2b4f
Change-Id: I9dd98d306ed2d26a859c93871b136cbf3ee6ec1d