[roll] Roll fuchsia [cobalt] Filter test projects out of the production registry.
A separate set of config data containing the full registry (with test
projects) is provided to the testapp.

Allowlist the cobalt_tests package to receive config_data. This is
not a new use of config_data, we are just providing the same set of
config_data for the Cobalt component, but using different data for
the test.

Original-Bug: 107294
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/715131
Original-Revision: 2df678d74458b8bed039bec786ef9f6bf6148781
GitOrigin-RevId: b83eb13e692d5568e4b73e6d20388ea02e372e16
Change-Id: I9e5e4812a99b1f4439a711423cbf4fa1498279cc