[roll] Roll fuchsia [starnix] Ensures HandleTable::insert_for_transaction ref integrity
HandleTable::insert_for_transaction can either returns a new handle with
a strong ref count of 1, or return an existring handle. In the second
case, before this CL, it didn't change the ref count of the returned
handle. As the caller was expecting to own a strong reference, it would
decrease it, and delete handles that were still referenced.

This CL updates the behaviour by ensuring that the strong ref count of a
returned handle is increased.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/715866
Original-Revision: 960b3eb46740c62812d0f50abe8cc3a8510ff32b
GitOrigin-RevId: 795c95fedb28a49f981fe19954af2adeaf692029
Change-Id: I75235bc2c0606f01b8c600de4da43f2d0509d107