[roll] Roll fuchsia [tpm] Add ExecuteCommand to the TPM Driver.
The existing ExecuteVendorCommand constructs the TPM header and
assumes a header will not be provided. This works well for non-session
commands but session based commands require tags to be written into
the header section of the TPM. This change introduces a new method
that allows the driver to accept a full command with a header. This
change also deprecates ExecuteVendorCommand and we should aim
to remove it in a future revision once the call sites have been
updated.

Original-Bug: 105028
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/709811
Original-Revision: 31ad0aee84346dd0df85885e395bfed72876ac9c
GitOrigin-RevId: 32dc1a3cb49d35e7de4faf9f0c3710ee0d5c8a61
Change-Id: I4dd95e7c340c8d5df4f743ef920fb31f74a99c71