[roll] Roll fuchsia [bt][fastpair] Use big-endian address in GATT notifications
https://developers.google.com/nearby/fast-pair/specifications/introduction#OctetOrder
notes that all multi-byte fields should be big endian.

Fix a bug where we are not sending the local address in big endian.

While here, add a warning if the received local address differs from
that of the local device. Make this failure non-fatal and mark with a
TODO for a future update.

Original-Bug: 107656
Test: fx test bt-fastpair-provider-tests
Test: Manual - verify via Wireshark that the outgoing GATT notification
has the address correctly formatted in Big-E.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/715727
Original-Revision: 35556b07ce22a00fbc5925de766985fb2610529f
GitOrigin-RevId: 382bfe58c6fe897f2eb50e90a57f10608e689857
Change-Id: I7c10dc03a5b830fee9f219a82ad466bb67752e6d