[roll] Roll fuchsia [ot-stack][lowpan-spinel-driver] Remove remnants from tree
This component is obsolete and we're not building it. If we keep it the
code will just rot, so we may as well remove it from the tree.

This lets us remove the defunct references to ot-stack.cmx

go/cmxtinction

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/715127
Original-Revision: 74e8fc5cecac3b806cb0a6640442302f3d5b3a0c
GitOrigin-RevId: a81fce1379a93ae1d36b25d453338bd9cdea73fc
Change-Id: I0c6e989ebc189acaa121bc3eb98e2108b38da8dc