[roll] Roll fuchsia [ffx][config] Context-awareness in ffx config (v3)
Re-landing this after major changes to lockfile handling to make it more
reliable and better able to deal with "unusual" circumstances.

This also removes (compared to earlier attempts to merge) the long-hold on
the environment lockfile that was intended to prevent overlapping
read/modify/write conflicts. That seemed to cause a bunch of unrelated
problems in lockfile handling that I believe have been largely fixed now,
but I want to land that in a separate change later anyways to isolate
its effect.

This is a reland of commit 21860e901fcd68fd37fea69cb10852adbcdf846a
Which was a reland of commit 4c113119d7a105b575544010aea20431abc46597
>
> Original change's description:
> > [ffx][config] Context-awareness in ffx config
> >
> > This makes it so that ffx config is aware of the context in which it's
> > being run, understanding to start three main contexts it can exist in:
> >
> > - In a fuchsia.git build tree
> > - Isolated for testing purposes
> > - Outside a build tree just being run on its own.
> >
> > This change doesn't really change any behaviour based on this
> > understanding, but it does do some extra cleanup of how environments are
> > constructed and managed.
> >

Multiply: ffx-e2e-without-target
Original-Bug: 106578
Original-Bug: 82518
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/711545
Original-Revision: 907185614d5289fbea730ba496779e2f3d5f0402
GitOrigin-RevId: 5a6de28fb20463d6d82f74b64f694fb7fd16508f
Change-Id: Ic00731c7dbf523ee24de38ddc8d18b8733652962