[roll] Roll fuchsia [blobfs] add query back to component runner test
When the fuchsia.fs.Query interface was removed, the substantive part of
this test was deleted, but it didn't get a replacement. Add the call
back into the test, this time with fuchsia.io.Directory/QueryFilesystem
instead, to make sure the request queueing is properly exercised.

Test: fx test blobfs-component-runner-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/686945
Original-Revision: 7e8327b6723de27702c29efc00445a20c4c40594
GitOrigin-RevId: 4a023e27aa59e7caa4af20894f4dc1af1fc48050
Change-Id: Ice0eaf77de72c59e123446209bf34dfce5b57f5e