[roll] Roll fuchsia [netstack3] Put sockets in accept queue only when it's SYN_RCVD
Previously we put a socket in `LISTEN` state into the SYN queue if the
incoming segment has and only has a SYN flag, it is guaranteed to
transition into SYN_RCVD but the logic is subtle. This change refactors
this and made it explicit that only SYN_RCVD sockets can go into the
queue. This change also makes it easier to install a retransmission
timer for the SYN-ACK we sent, which will be done in the next commit.

Original-Fixed: 107592

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/715279
Original-Revision: df00f648c0523ffbc1239a043cef4cf8131e4692
GitOrigin-RevId: 2c34610455be97453ff7bed6f31f21079754d600
Change-Id: Ifaacb8cf587627e0203962b5710b8ba45170aacb