[roll] Roll fuchsia [golden_go_proto] Ignore comment whitespace
The diff is sensitive to whitespace changes in comments, which makes
it more difficult to roll Go versions due to changes in the formatting
of generated Go code for protos. This CL changes the script to remove
any whitespace in comments before comparing.

Also updates the error message such that the suggested `diff`
invocation no longer ignores whitespace, since this was causing `diff`
to return nothing in cases where the golden verification script was
finding differences. (Now, `diff` is at least as sensitive as the
verification script, which is what we want.)

Original-Bug: 93043
Test: verify_golden_go_proto_file_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/716153
Original-Revision: ac474942fee713771f3162fbfe1f17513aab1f2b
GitOrigin-RevId: 128aa722483998775d97bb3dc6353d6ae1e3aa31
Change-Id: I96227afc80eb9fe11dd6f769bbbe296f786614d0