[roll] Roll fuchsia [dfv2][compat] Remove deps on VFS
The compat driver is using a VFS to host devfs nodes.
The compat library shouldn't do this, because the VFS
library shouldn't be exposed in the SDK.

Simplify some of the compat library logic in the process.

Original-Bug: 107384
Multiply: driver-compat-test

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/716014
Original-Revision: f9cfbbfb5dbb986cc0db36e63a93828d631153eb
GitOrigin-RevId: 764053f77f110f48a4d1d6e12fbfcee42637f52e
Change-Id: I6f2cafacec70a5de15ec952fa8c4a904ce9dab14