[roll] Roll fuchsia [component_manager] Append `OpenFlags::Directory` when opening directory capabilities
There is a bug with routing subdirectories of `/dev` to components.
Components would not be able to open these capabilities, experiencing
PEER_CLOSED errors or other vague errors like `Broken Pipe`.

When opening subdirectories of `/dev`, devfs cares about whether
the `OpenFlags::Directory` flag is passed in. This flag determines
whether the subdirectory should be opened as a directory or service.

Component Manager did not pass this flag, so devfs would attempt
to open the subdirectories as a service, which means they would
not respond to `fuchsia.io` calls (if they opened at all).

It is worth noting that Component Manager does pass in
`MODE_TYPE_DIRECTORY`, but this looks to be ignored by devfs.

This CL ensures that the `OpenFlags::Directory` flag is now passed
in when opening any directory/storage capability, adding stronger
guarantees about what type of capability Component Manager is asking
for from the component.

Original-Bug: 107862, 106028
Run-All-Tests: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/716503
Original-Revision: 45a21c2eea9d2838715bf35f825ff2da3ebac809
GitOrigin-RevId: 753526415cb510e1df01d6196aa4aaa7f3f7b508
Change-Id: Ifd4c1b99218ef2251bb9d2711714f836810ce3a7