[roll] Roll fuchsia [ui][input-pipeline] add touchpad event data to inspect
Add a log of recent touchpad events to the inspect data exported
by the input pipeline.

Along the way:
* update the `main()` function in scene manager to increase the
  size of the Inspect VMO
* add a `stats` node to the Inspect VMO, which will serve as
  a hint that the touchpad data was truncated

Design note
-----------
This implementation logs touchpad data after the data has been
parsed from `touch_binding::TouchpadEvent` into
`gesture_arena::TouchpadEvent`.

This approach is "lossy" in two respects:
1. Some events are rejected during the conversion (e.g. events
   without a touchpad descriptor), and such events will not be
   included in the inspect data.
2. The converted event contains movement in millimeters, rather
   than whatever units were provided by the device driver. The
   original information on units cannot be recovered, nor can
   we detect errors in the unit conversion.

Additionally, the inspect schema for touchpad events is lossy:
* The inspect schema represents contacts as a set indexed by
  contact ID.
* This contrasts with the data from the driver, which uses a
  vector of contacts, each containing an ID.
Hence, if the driver repeats a contact ID within an event,
and provides inconsistent data for the repeated contact(s),
the inspect data will not accurately reflect the driver data.

I think this is a reasonable choice, as I think the best
long-term direction for dealing with ambiguous data is to
normalize said data before it is processed by the gesture
arena.

But we can revisit this later if needed.

Performance note
----------------
Per advice from the feedback team, it's important to ensure
that the touchpad data doesn't greatly increase the size of
a feedback report.

To evaluate the impact of this change, I did the following:

* Installed this patch on an Atlas m3.
* Moved two fingers on the touchpad for ~15 seconds,
  to fill the event log (125 Hz * 15 seconds = 1500
  events, which is greater than the limit of 1250 events).
$ fx iquery show core/ui/scene_manager > events.json
$ grep position_y events.json | wc -l
  2498
$ zip events.zip events.json
$ wc -c events.zip
   40904 events.zip

Original-Bug: 99334
Test: fx test input_pipeline_lib_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/715741
Original-Revision: 7dc013cfea831ae01eee740d9ba9e8a65a57b340
GitOrigin-RevId: 0d951d7e3f2984bddb6abf1ee435c1b95f9c8c88
Change-Id: Idc697b80494f06fe368cb71cea8aa284b1b2a6fe