[roll] Roll fuchsia [build] Allow goldens to be reformatted before diffing
It can be useful to have goldens less sensitive to formatting
differences (e.g., to prevent breakage in the case where a new version
of the formatter is being rolled into the project). This was also
a feature of `golden_test()` desired for `golden_files()`.

As a proof-of-concept, we migrate the libfuzzer golden diffing over to
using `golden_files()` with the formatting specification.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/715738
Original-Revision: cba0cef772d281f1800392a47171b0e8b76d20d2
GitOrigin-RevId: 01b317e709673f14352a7ba40fe02c1aeb83b5fa
Change-Id: Ia0d17d171b5768ff570391288589a614d8653392