[roll] Roll fuchsia [performance] Remove Insntrace Binary
Insntrace isn't used as far as I'm aware, and the actual implementation
is experimental and Intel only. In addition, between the garnet
migration and cfv2 migration, and now a ktrace migration, it's been
using up time to keep it alive for no benefit.

This removes the insntrace component that interfaces with the driver,
which is up next to be removed.

Original-Bug: 107507
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/715512
Original-Revision: fa1a61e781bea0b71474be1fad612475ad07c7d0
GitOrigin-RevId: f5ab623695b9cb42889cb501fcc3c542faaebfbb
Change-Id: I4456ce32bb1dc2d6ddaec818821ea38052276d95