[roll] Roll fuchsia [kernel][vm] Option to place anonymous reclaim queue
Adds an option to place anonymous pages in the reclaim queue and
support for aging them. With this option being a compile time
constant, none of the added code presently exists in final builds
and is effectively a no-op.

This option is currently harded coded to false due to both lack of
tests, and the ASSERT syscalls/object.cc that would trip if enabled.

Original-Bug: 60238
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/715786
Original-Revision: 56e39ec778b960ed846ac87d2bead89810c98913
GitOrigin-RevId: aa19712a2bfb5a95431861464a681447f51d877b
Change-Id: Id98ddb0906b1c17ee7bbc4011ba6eacaa03d98dc