[roll] Roll infra/recipes [auto_roller] Add "GitWatcher: ignore" footer
With cl/469276783 GitWatcher supports completely ignoring CLs with
"GitWatcher: ignore" commit message footers. This is an issue with
single-commit roll CLs that include a reference to a bug in the main
body of the original commit message, like
https://pigweed-review.git.corp.google.com/107390/ (see the associated
bug for details).

This is usually more of an issue for Buganizer bugs than Monorail bugs
because patterns for matching Monorail bugs are much more narrow, but it
could affect Monorail bugs too. Keep it behind an option so projects can
choose to use it or not.

Original-Bug: b/243177456
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/715091
Original-Revision: 31840dacdf4069127db151b7fa2a1a7f213d625a
GitOrigin-RevId: aa6564f914ea998edbb1c80d2c4189b18a23ff84
Change-Id: I25e18d45af17647da5174ecc74e053afea0e31ac