[roll] Roll fuchsia [component_manager] Bind routing tasks to component instance
Instead of binding routing tasks to the namespace's lifetime, bind them
to the component instance instead.

This reduces the probably that a message sent to a component's namespace
at the end of its execution gets dropped because the deliver of the
request raced with component teardown.

This is a fix sufficient to address flakes that emerged with the
migration to event_streams (fxbug.dev/107764). However, it is not
bulletproof. If the component is destroyed it will still cancel the
task, and it's still possible (albeit much less likely) for a routing
request to be dropped if the namespace VFS hasn't yet processed the
request from the channel.

Followup work should be done to investigate a more complete and robust
solution. For example, we may want to wait for all tasks to terminate,
possibly in combination with a timeout.

Original-Bug: 76579
Original-Bug: 81980
Original-Bug: 107764

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/716168
Original-Revision: 436572124bae5288e0f6dbe2a4456346b0f3d04e
GitOrigin-RevId: 20ca4f5fe852e740f1ea0c6d40991f0ba050c858
Change-Id: Ib0d9195a8865781bd4bb1997fe311e8e67219ec3