[roll] Roll fuchsia [zither] Fix _golden_test FIDL for new go/format
zither_golden_test is sensitive to formatting changes in generated Go
source, which is blocking rolling Go to 1.19 due to changes in
go/format. As a workaround, this adjusts the FIDL source such that the
generated Go is formatted the same way before and after 1.19.

Additionally, this change adds support for ignoring whitespace changes
to golden_files, as the new go/format has different behavior not only
with respect to blank comment lines but also with respect to tabs vs.
spaces in comment indentation.

Original-Bug: 93043
Original-Bug: 107996
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/717049
Original-Revision: 2975d6bc8c24701207fe03b1313d206d05bf2b64
GitOrigin-RevId: 3d5ab5fb73bb493d7c5b84d94c1c568922d6f343
Change-Id: I4adfd5d5c437603c8f76ac925d2ed91c90f4dad1