[roll] Roll fuchsia [syscall][pci] Avoid a potential overflow.
When verifying the size of the PCI window descriptions in
sys_pci_init, be careful how we perform the calculation of how large
the flexible portion of the structure is.  This calculation depends on
a value passed by the user in the structure itself which cannot be
fully trusted.

So:
++ Change the multiplication used here to one which tests for overflow
   instead.
++ Change the addition which could hypothetically overflow into a
   subtraction operation which cannot underflow (because of a previous
   check).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/716818
Original-Revision: c0a35647976da7d34cc58a081698771192c7d704
GitOrigin-RevId: a0227da8a9370999340eed090d2941cb7e84883f
Change-Id: Iccc23c5164b9e29bc0be7d2b16654e837258f76f