[roll] Roll fuchsia [performance] Remove insntrace from cpu-trace driver
Insntrace isn't used, and the actual implementation is experimental and
Intel only. This removes the the ability for the cpu-trace driver to
connect to the insntrace fidl protocol making it solely a pmu driver.

The fidl protocol itself will be removed in a following patch.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/715521
Original-Revision: 34980f771abb9c6b33cf5a2c442ea3b0e6a169e9
GitOrigin-RevId: 9614ca77b226576a529ba17d0157c92a61d9a39c
Change-Id: I3d7a46598fddf0e7154ebc706c2c714f20035c8b