[roll] Roll fuchsia [coverage] Disable hwasan globals on clang_coverage_test_bin
This particular target fails to link because a tagged version of
__llvm_profile_counter_bias provided in clang_coverage_test.cc causes
an overflow check to fail when resolving a relocation in an
unistrumented profile runtime. For now, we can just disable
hwasan-globals for this specific target. This is done through a config
so we guarantee the flag comes at the very end of the compile
invocation.

Original-Bug: 107896
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/716632
Original-Revision: 2c9ff0521c2f93220869c4566706bae0c8eb5e96
GitOrigin-RevId: 56f4b0e3c221bbd1b018e769977cba253e6c2d54
Change-Id: I0445c2dfaf6f3622c230708b3083b84010467814