[roll] Roll fuchsia [element_manager]: don't drop element til closed
ElementManager::run_element_until_closed() blocks until the element is
closed. This function is called as an async task, and the function
(and therefore the task) takes ownership of the `element` instance.
This way, the `element` instance is not dropped until it is considered
closed.

For elements based on CFv2 components, dropping the element does not
appear to cause an issue. But elements for CFv1 components hold onto an
`App`, and the `App` holds a `controller`, and the `controller` "Keeps
the component alive until `App` is dropped."

A recent change removed the `element` from the call to
`run_element_until_closed()` because the function no longer needed the
`element`. Tests against components migrated to CFv2 still worked. But
there is an end-to-end test that has not yet been migrated, and the test
now fails because the `element` is no longer passed into the task.

This CL reverts the part of the change that removed the element, so the
element will no longer be dropped until the function completes.

Original-Fixed: 108032
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/717362
Original-Revision: 5535ee25a96eeb30ca7c23ce833a83e497960480
GitOrigin-RevId: d18e9deb75ed6ba59c43a48abb4da48294fab4fd
Change-Id: I0b2fca6cfa5e6d65b875f6db090fa46154bb0c1d