[roll] Roll fuchsia [fxfs] Fix volume termination race
It was possible to race closing of a file against locking a store.  If
the file was unlinked, this could trigger a tombstone which would trip
assertions if the store was locked.

This CL fixes the issue by ensuring the pager is properly terminated
before attempting to lock the store.

This CL also:

  + adds new lock states so that we can more easily diagnose these
    issues in future.

  + ensures that we do the initial reap of encrypted object stores.
    This was missing.

  + allows unlocking volumes when the filesystem is read-only.

Original-Fixed: 107470
Original-Fixed: 107695
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/715789
Original-Revision: eb31c994cfd5474a472602d7470a6c44195bf565
GitOrigin-RevId: 85e563c6b8d4e0513fc0fe2c2a46e34398d2b3b7
Change-Id: I30eba680a0979fdfdb95cacb976653d72558f0b2