[roll] Roll fuchsia [driver_runtime] Protocol discovery - call service_connect op.
This is a temporary implementation using v1 of the protocol discovery
API. The child driver tries to connect to a driver runtime protocol
using DdkServiceConnect, however in the compat-shim this tries to
make a FIDL RuntimeConnector::Connect call to the parent. We need
to implement this protocol if the parent driver is running the compat
shim.

This is so the wlan driver can work in compat mode in DFv2. We should
replace the API with the newer protocol discovery when
available.

Original-Bug: 107866
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/716819
Original-Revision: 732cd6c145afa2eaf9e4ca92777121cae831c61f
GitOrigin-RevId: be64e44617e7608d463a78386654bf904bce87a1
Change-Id: Idbd7e6f4b3d69bf0f04b59e9268f587a4423d448