[roll] Roll fuchsia [cts] Impl more policies for JSON schema changes.
Changes to the value of an "enum" key are now classified as breaking
changes. Changing the value of "additionalProperties" to True is
non-breaking, but to False is breaking. Any breaking change that doesn't
come with a change to the "id" field is another error.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/715689
Original-Revision: c7de156714a539c550de1e645939b76d3ea7e612
GitOrigin-RevId: 07f48867d2ae8493e55c59a0a7daa2985d71f6a2
Change-Id: I67e9cb84a80b90cb6402b62239115186e288dd04