[roll] Roll fuchsia [netstack3] Move create listener into datagram.rs
Move the functionality of creating listener sockets into the datagram
module and delegate to that when allocating a UDP listener.

This also has the advantage of removing the linear scan that was being
performed when allocating a socket for a UDP listener since the same
information is available in constant time from the socketmap.

Original-Bug: 104462
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/714620
Original-Revision: 58ea6f398fa20ccb66cd664af549f194ea1a4e52
GitOrigin-RevId: bfe04fe2cab51fa6a5a261024cbb31f918826199
Change-Id: Id4e74306673ae1d0da31922841896fed0986cf39