[roll] Roll fuchsia [nxpfmac] Create an event handler
In order to handle events from mlan we need a centralized event handler
that allows for registering of event callbacks. This creates a very
basic event handler that allows for registering event callbacks per
interface or for all interfaces or events that are not associated with
an interface.

Test: Included unit test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/708983
Original-Revision: 0ae6508de091d0838116b42c70b6ef277f426989
GitOrigin-RevId: f559b9f89428051837cb9ce43e980aa5833a8459
Change-Id: I310781bc4185f713500438fd50a33b57ec32eda2