[roll] Roll fuchsia [virtualization][vsock] Run the util in its own thread
At the moment we're running the util in the main thread and the test in
a new thread which is explicitly disallowed (can't pass FIDL pointers
between threads).

This swaps to running the util in a new thread. This is a bandaid to the
larger issue of this test harness not being thread safe to unblock
some vsock work.

We'll need to re-visit allowing non-blocking commands in the future.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/717591
Original-Revision: eac1f5ea9f71136cf5e90425b6191e0ecfad8cdd
GitOrigin-RevId: 7d9116df7701220b350701507e59b7800f23cb0a
Change-Id: Ie5f2bb8f68b390b589cd378d60e051a34aec1541