[roll] Roll fuchsia [settings] Add inspect structure for external apis
To log external service events to inspect, the service context's
ExternalServiceProxy will be used to handle ExternalServiceEvents from
a common point. This introduces the event and structure for doing so,
and converts direct calls on the ExternalServiceProxy to use the call
and call_async macros, in order to log the external request.

Original-Bug: 94361
Test: fx test setui_service_tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/705626
Original-Revision: 984692a2dc477f907951408f0102efa8b237301f
GitOrigin-RevId: ff13d82e726b5bf97b0330b6221de6b21b9851fb
Change-Id: I629085d1ab7dd07970f6e2c5ff83bf58cf8f3ff5