[roll] Roll fuchsia [virtualization] Merge linux_runner with termina_guest_manager
This updates the code in //src/virtualization/bin/linux_runner to link
with the //src/virtualization/bin/guest_manager code. We then use a
pointer to the GuestManager from the linux_runner instead of connecting
over FIDL.

This is just a mechanical change and no improvements are added yet. For
example, starting the vm using the TerminaGuestManager protocol does not
yet wire up the needed vsock services, but we will add these
improvements in a following CL. We also don't yet move or rename the
linux_runner code to minimize the noise in the CL.

The effect of this CL is we produce the component
'termina_guest_manager.cm' out of //src/virtualization/bin/linux_runner
and delete the former, standalone termina_guest_manager package.

Original-Bug: 107399
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/710344
Original-Revision: 907cd7a24652ec41f89f57b682e3a910c2ac273b
GitOrigin-RevId: afe49d2b916d2f25204eac6ec05eceed38a2b659
Change-Id: Ib13705438a6bd112e0bb12f9a2d897039522f60f