[roll] Roll fuchsia [lowpan-ot-driver] Reset OpenThread stack if busy for longer than 30 sec
Scans are somewhat blocking operations---the device cannot actively
participate on the network while one is in progress. Occasionally we can
run into bugs like <fxbug.dev/106509>, where the scan never finishes.
Because there is no way to cancel an ongoing scan in OpenThread, the
only way to get ourselves out of this state is to reset OpenThread.

This CL introduces a `scan_watchdog` which monitors for when the scan
starts and then makes sure that it lasts no longer than
`SCAN_WATCHDOG_TIMEOUT`. If it does, it will terminate the loop, which
will cause OpenThread to be reset.

Original-Bug: 108271
Original-Bug: 106509
Test: Manual verification.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/718442
Original-Revision: 883cdde982be4cb8f847fa54a7e2cd852d776fe1
GitOrigin-RevId: 26c0b7bb01eca28ff71fa0f8b268ee8212405b68
Change-Id: I2ec8b5b6223214fc253714f2786f277b372cb6f5