[roll] Roll fuchsia [fuchsia-tpm] Implements fuchsia.tpm namespace.
This library intends to implement the fuchsia.tpm namespace
so that it can be shared between the cr50_agent and the tpm_agent.
This allows variation by board type where the cr50_agent is a superset
of the functionality in the tpm_agent. By placing this code in a shared
location it prevents code duplication / drift between the two components.
This initial implementation implements just transmit. Both deprovision
and provision first require our TSS implementation to be ported to Rust.

Original-Bug: 10528

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/712387
Original-Revision: 682454b7b170b7b6fea4d2b8161c8da25120cfb6
GitOrigin-RevId: e9bd06755a891eaeb3841631a7511c2bcd6320d9
Change-Id: I2b6a87a3fc281ed2dfe248a017cf61986e3cd849