[roll] Roll fuchsia [tpm-agent] Add support for fuchsia.tpm.Command.
This change adds the groundwork for piping the three new protocols
into the tpm_agent: Command, Provision & Deprovision. On launch
the tpm_agent will scan for a valid TPM20. If no vaild
TPM20 is found the component will terminate. If a TPM20 is found it
will start a fidl server services the three protocols. Currently
only Transmit is implemented with Provision and Deprovision
returning not implemented if called. tpm_agent uses the
library fuchsia-tpm to implement the service so it can share
code with cr50_agent.

Original-Bug: 10528
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/712989
Original-Revision: 87c2931695515504bce80eb1d43ea9b664b76f51
GitOrigin-RevId: 08d83464ba3035866240a8d6c2c18a57eabad7e0
Change-Id: I64fdae28045567ac36411649495cda1cbc4656b0