[roll] Roll fuchsia [testrunner] Properly exit package resolving loop if task finished.
Before, we'd wait for the function to finish but it actually goes
through all the package urls and does a best effort to resolve, meaning
it doesn't exit if one fails. In the case that it fails due to a
connection issue, we try to reconnect for 2 minutes and we repeat that
for each url. While the loop does break if the context gets canceled,
the context doesn't get canceled until the function returns or we hit
the botanist timeout. So we should have a separate context for the
package resolving function and cancel it when execute() is finished.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/718007
Original-Revision: 10fc3ec658fd9b87bff5e1b8e0de7c59d2af39cc
GitOrigin-RevId: 9652e351a1a5c16980499a9ca1f3c30942f98a2e
Change-Id: I65cd2cf85a64f1115443d900b9169769642862f7