[roll] Roll fuchsia [virtualization] Use structured config in termina_guest_manager
There are a number of configuration options to the termina_guest_manager
that are augmented by setting preprocessor symbols and recompiling the
target sources. This is difficult to test since code may be #ifdef'd out
and won't always compile.

This CL introduces some structured configuration[1] to
termina_guest_manager so that we can still control behavior using GN
args, but we also remove the need to #ifdef code. Additionally
structured configurations are much easier to change within test
contexts.

This CL also removes the vestegial support for using a prebuilt stateful
partition. This was useful when network support was not reliable enough
to bootstrap the container but hasn't been used in recent months.

Original-Bug: 107399
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/714871
Original-Revision: cc8d451d3031b7297bee97068286998aa73db7cd
GitOrigin-RevId: f38bdb71f5ad8c2771e3c0a87f2e0cab3ae4cc00
Change-Id: I76b3c830efa4ef85949b0e159556ec229a8c7e90