[roll] Roll fuchsia [virtualization] More explicit guest configuration
This migrates more configuration to be explicitly stated by the guest
packages instead of having global defaults.

Currently, we have opinionated logic for configuration in:
  1) The guest CLI, where defaults are enforced for things like which
     devices are present, or the default amount of memory or CPUs.
  2) The GuestManager, which has logic in the config parser to also
     enforce some defaults for the same things in (1).
  3) The linux_runner/termina_guest_manager, which needs to add an
     override in C++ code to remove the virtio-gpu.
  4) Integration tests where we restate some things (like no virtio-gpu
     for termina).

This makes the system difficult to reason about and it can be confusing
to track down where configuration is applied.

To remedy this, we start adding more static configuration to the guest
paackages themselves. We do this by extending the existing GN templates
to add support for all the devices a guest may use. The idea is to move
defaults from the CLI (which are global and apply to all guests) and
into the guest packages.

In following CLs, we will begin to remove the default logic from the
different argument and config parsers and defer to the guest manager
components to be the primary source for building the guest
configuration. Then, the CLI will only supply values that override or
otherwise augment the guest-specific defaults as supplied by the guest
package and guest manager.

Original-Bug: 107399
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/711425
Original-Revision: f99950a261cd823709219429eca421b818fb73d2
GitOrigin-RevId: a008ecfecae983fd495f141df974bca66cf09df0
Change-Id: I80d96766b79649e59b6310cef1d1b65a0139d5dd