[roll] Roll fuchsia [virtualization] Stop providing default values in the CLI.
Instead we want to let the target guest package determine the defaults
for these. Specifically consider an option sucha as virtio-gpu, which was
previously a 'default on' option. If we launch termina using the CLI, we
do not want there to be a GPU device added automatically by the CLI.

In other words, the CLI would not mutate any option of a guest package
unless it is specifically requested as an override.

Original-Bug: 107399
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/711426
Original-Revision: 7c12ae914506e74cef9ce34c3a5e0a024aef5df5
GitOrigin-RevId: b6f1969031516265dae10b7627812e775f60fc60
Change-Id: Ida0ef52d11452bd4aca4a73489bbf46d702582a0