[roll] Roll fuchsia [starnix] Always handle special errnos in dequeue_signal
Syscalls can return special errno value to let the kernel restart them
in case of interruption.

Starnix handled these in dequeue_signal where is replaces the rax with
either the EINTR errno value, or with the right value to ensure a
restart. But it only does this if a signal is queued. Otherwise, it let
the errno value as it is, which is incorrect, as user space should never
see these values.

This CL updates dequeue_signal to always handle the special values,
whether there is a signal queued or not.

Original-Fixed: 107926

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/718843
Original-Revision: 72e42a1087db932b6a9745143924f2f7cc4c99ea
GitOrigin-RevId: d331fe1aad405fa8e91018a97d3aee810b78979a
Change-Id: I06ff46055949b3ad55b0ea1869b052941bfec385