[roll] Roll fuchsia [bt][avrcp][a2dp] Subscribe to player changed notification events.
When an a2dp sink, we should detect when the available players have
changed, query to find the new players to see if there are any that are
active. If none of the players are active, we change the player state to
idle.

We also should detect when the addressed player has changed. This event
will be treated similar to track change event and should trigger
attribute update.

Original-Bug: 103847
Test: fx test -v -o bt-avrcp-tests && fx test -v -o bt-a2dp-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/703168
Original-Revision: 2784cf59e928336a149139cd11b525752d1c7b6a
GitOrigin-RevId: e03509b77dd38610d5ff105a397c3015f9c22566
Change-Id: Ic093542bccb5a69ff237604ef6215d516a522049