[roll] Roll fuchsia [minfs] Fix shutdown race with periodic sync.
If the periodic sync is in the middle of running during destruction,
there can be a lot of havoc. This will now try to dequeue any waiting
sync calls, and if it can't find one to dequeue, it ensures that one
isn't running by pushing a task into the queue and blocking on it.

Removes vestigial RefCounted interface and marks minfs final so that
we have full control over destructor ordering.

Original-Bug: 107070
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/714069
Original-Revision: 443412ec4546ef7f21d71a3c58b7a0d5bbf02a3a
GitOrigin-RevId: bd11fcca779aa7576ea0cfc5cc285d0922cc10f9
Change-Id: I31a5f62ed438c410e6691b7499408d127a98e0ea