[roll] Roll fuchsia [fuchsia-tcti] Add a Fuchsia TCTI Interface.
The TCG TSS 2.0 TPM Command Transmission Interface (TCTI) is the
lowest layer in the TPM Software Stack (TSS). It is a trivial
interface that defines how bytes are sent to and recieved from
the TPM. This API must be C compatible as its lifetime and
calling structure is managed by the TSS 2.0 library which we
imported in an external repository. This CL simply creates
a small internal context which opens a FIDL connection
to the tpm_agent or cr50_agent depending on board type
and uses the fuchsia.tpm.transmit.TransmitCommand interface
to send  commands to the TPM. Responses are recorded on
the heap in a recv_buffer and data is removed from that
buffer as calls occur to fuchsia_tpm_recv.

Original-Bug: 105028
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/714016
Original-Revision: 21a02e5bffbad465dcce591b68dcae58e272127a
GitOrigin-RevId: a903ff006370a3da8e96604d888043807bd4411b
Change-Id: I3e09300c4ef1981ffe2dd40462132dbf1fd65710