[roll] Roll fuchsia [virtualization] Unify termina launch methods
TerminaGuestManager exposes a lower-level GuestManager::LaunchGuest API
that allows for a lot of configuration options. This is the API that is
used when launching a guest from the `guest` CLI (ex: guest launch
termina`). We also have the LinuxManager FIDL, which is used by tools
like vsh that provide feedback that is specific to the termina guest.

Before this CL, launching termina with `guest launch termina` would
produce a different result from launching with `vsh -c` because the
vsock listeners and stateful block device would not be included when
launching with `guest launch ...`.

This CL refactors the GuestManager & TerminaGuestManager so that we
always launch the guest correctly, regardless of the interface used.

The TerminaEnclosedGuest integration tests will now include all the
logic in TerminaGuestManager. This allows us to remove the test-specific
startup listener and instead update the test to listen to updates from
the LinuxManager FIDL. Since these tests do not currently use
containers, we use a variant of the termina_guest_manager package that
includes a structured config that causes the termina_guest_manager to
stop before the container system is started.

Original-Bug: 107399
Test: `guest launch termina` now behaves similarly to `vsh -c`. Using
      `guest launch` will provide low level boot spew instead of the
      progress indicator that vsh has.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/714628
Original-Revision: 016fd31b8f7a60ad43b8d38ec11b81ac53c3321e
GitOrigin-RevId: 84ae960db3b02e84dd148ccff091c2dcb3481c37
Change-Id: I69601252f3017b539bd016b113a477403914b23e