[roll] Roll fuchsia [fidl][cpp] Make fidl::WireRequest just an alias
Now that we have predictable, somewhat sane ways to name the types that
represent method arguments we can stop using fidl::WireRequest<M> to be
a special type for requests - we can just use the actual underlying
type.

It can still be useful to have a way to get the request type for a given
method so this leaves fidl::WireRequest as that, but it's just an alias
based on some special generated traits stuff.

Original-Bug: 101525
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/718644
Original-Revision: a25adbcae586ac3c84666efeae0ec6df968e2051
GitOrigin-RevId: 328e972247f9a1782d2ae25e1109cc9d7caac7d6
Change-Id: I9c3260a6cc5e9a2a999ab410a4686e7ed19bb3e1