[roll] Roll fuchsia [ffx] Improve clarity of the plugins file template
Rewrites the logic blocks in the plugins.md file to be easier to
comprehend, with fewer layers of indirect logic and clear exit points to
the generated functions.

This also renames the template files to not imply that they're
documentation, but rust jinja templates. All references I could find to
them were also changed. I'm not sure why they were given .md extensions
to begin with, but now editors will highlight them properly and cq
(hopefully) won't treat them as no-ops for CI purposes.

Note that this reveals a potential flaw in the `ffx schema` output that
might need to be resolved, as replicating the prior logic results in
some output being masked (a ProfileMemoryOutput on `ffx profile memory`
is not included with this logic, but would be with logic that didn't
only print out the plugin's output if there are no sub-plugins.

However, it's not clear how that should be output, the format of this
does not lend itself to adding leaf items under branch nodes.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/718424
Original-Revision: 661ae9eed5967ad5ff8c3bb3864aca01f2cb66aa
GitOrigin-RevId: c8ae8fbad04749318c545eae15419fa1c73750a4
Change-Id: Ie8980b7a1553bd71331c2b14bf9a8fc37880a156