[roll] Roll fuchsia [debugger] Redirect debug_agent's log to zxdb
Provides a general mechanism to allow the debug_agent to display logs in
zxdb. This is required by the TestLauncher because the error is only
available asynchronously. It's also useful to display info and debug
logs in one place for debugger development in the future.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/716803
Original-Revision: b6f06d9546812484e90f78af5564956c5db8aed2
GitOrigin-RevId: 366c3d819b7ec012b326e57cd193553bf96c26cb
Change-Id: I882ab59f41faf9ea250dd4df24ec074fd6b6c346