[roll] Roll fuchsia [kernel][time] Save timer offset on suspend.
This ensures we do not get time goes backwards errors on resume. It only
affects the get_ticks syscall. This is a short term measure so that we
can test suspend and resume. In the longer term a revamp of the vDSO is
planned so that ticks can be offset in a more holistic manner.

Original-Bug: 89735
Original-Bug: 91701

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/698004
Original-Revision: 6774c2a4e4a49d6d430abfa35283bb9510d35b46
GitOrigin-RevId: c998200c090c3d4a3a8b95bc8592d9dc666a0151
Change-Id: I9effff03cb9120290feab75cd05daf03e7c433e0