[roll] Roll fuchsia [microbenchmarks] Add same-CPU and different-CPU variants of a round-trip test
Measure the IPC round trip time when the threads involved are pinned
to the same CPU or different CPUs.

The motivation is to investigate how factors like lock contention and
IPI latency affect the different-CPU case compared with the same-CPU
case.

Add the plumbing for making this work for both the single-process and
multi-process cases.  The plumbing for the thread affinities is a bit
fiddly.  The parent thread affinity needs to be set via a wrapper for
the test, while the child thread affinity needs to be passed through
the test into the new process or thread.

I checked manually, by looking at traces, that the threads do actually
get scheduled on the same or different CPUs.

Original-Bug: 45802
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/718443
Original-Revision: aae56020d6a61e8794db662601caedfb798fa263
GitOrigin-RevId: 4d00ce66d8951d6029e391c0c884499b1bf74548
Change-Id: I1ad82a64ae4d74875c05e82b4ed88f76393fc4e1