[roll] Roll fuchsia [hypervisor][x86] Disable interrupts until SetMigrateFn
We need to ensure that the `AutoVmcs` is created before `vmcs_init` and
lives until `SetMigrateFn` is called, as it disables interrupts and
prevents us from being migrated to a different CPU before the thread
migration function is set.

Previously, we could be migrated between `vmcs_init` and `SetMigrateFn`,
and therefore would not correctly restore the VCPU state on the new CPU.

Additionally, make sure that we `vmclear` while interrupts are disabled,
for exactly the same CPU migration issue.

Original-Bug: 95763
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/719206
Original-Revision: 2d8797843368ccfd654ee5c7084add1ae19cf4f6
GitOrigin-RevId: 2ca9f22b3f7b98131ace359180f88a989e208916
Change-Id: Ife1cc175e4d03e0af334515eac98590e7f8092a6