[roll] Roll fuchsia [build][fidl] Allowlist HLCPP without _hlcpp suffix
This CL adds an allowlist of targets that can depend on HLCPP bindings
without using the explicit "_hlcpp" suffix. The allowlist is very broad
but will be narrowed in later CLs.

This also starts out by adding the suffix in a few places that are hard
to migrate mechanically, e.g. in templates or where the label is
constructed by string manipulation.

As a side-effect, this fixes some cases (measure-tape,
dangerous_identifiers) where we were incorrectly depending on HLCPP
bindings when the JSON IR was all that was needed.

Original-Bug: 108354
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/718410
Original-Revision: 4dae4ac485974d9886adb0704efb49d1fb445b12
GitOrigin-RevId: e4b1b423422159a981f009af4cd463a756781156
Change-Id: I0ba084e38abc25613419b4905357e2cce34694b2